==> scripts/04. CFT_Run_add fertiliser application method.Rmd <==
---
title: "Run CFT"
author: "Alexa Varah"
date: "02/10/2020"
output: html_document
fontsize: 11pt
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


The Cool Farm Tool (CFT) is an Excel workbook which enables farmers to estimate their carbon emissions.  
It's designed to be used on one field or farm at a time. Data entry takes quite a while.  
This script takes field management data (provided in an Excel spreadsheet) for multiple fields/farms and enters it one field/farm at a time into the CFT.  
Carbon emissions for each field are calculated in the CFT workbook, then the R script extracts the required emissions data.

# Load packages
```{r load-packages, message=FALSE}
rm(list = ls(all = TRUE)) # make sure envt is clear
library(readxl)
library(tidyverse)
library("XLConnect")
#library(tinytex) # for creating pdf markdown doc
#options(java.parameters = "-Xmx1024m") #Can set java.parameters = "-Xmx1024m" as this allocates 2GB of RAM to rJava if using a large dataset
```

# Load data
I'm loading some farm management strategies which I've run through the CFT manually and so it has some results columns, which I remove.  
Make sure to include 'trim_ws=FALSE' because the CFT has white space in one of the input variables so this needs to be retained in the input file.
```{r load-data-&-CFT}
# Farm management input data
#mgmt_data <- read_xlsx("../data/farm_mgmt_data.xlsx",1,trim_ws = FALSE) %>% 
#  select(-contains("results"))

mgmt_data <- read.csv("../output/cft_input_tidy.csv", header=TRUE)

# Cool Farm Tool emissions calculator
CFT <- XLConnect::loadWorkbook("../data/CoolFarmTool.xlsm") 
setStyleAction(CFT, XLC$"STYLE_ACTION.NONE") # Keeps the formatting of the original document
```

# Specify generic data 
We need to specify the country and units from drop-down cells in worksheet '1.GeneralInfo' in the CFT workbook.  
```{r specify-country-&-units}
# Country
writeWorksheet(CFT," - United Kingdom", 2, 7, 5, header = FALSE) # sheet number, row number, column number respectively

# Units
writeWorksheet(CFT,"Metric", 2, 8, 5, header = FALSE) #'header = FALSE' stops weird placement
writeWorksheet(CFT,"hectares", 2, 12, 6, header = FALSE)
writeWorksheet(CFT,"tonnes", 2, 14, 6, header = FALSE)
```

# Create lookup table
Create a lookup table specifying where each piece of management input data should be sent to in the CFT workbook.
```{r make-lookup-table}

cols_to_use <- mgmt_data[
  which(colnames(mgmt_data)=="fieldname"):
    which(colnames(mgmt_data)=="spraydays_BGgly")
  ]

lookup = setNames( 
  data.frame(matrix(data = NA,
                    nrow = ncol(cols_to_use),
                    ncol = 3)), 
  c("DestSheet","DestCol","DestRow"))

row.names(lookup) = names(cols_to_use)

# column in CFTdata:              1   2    3   4    5    6   7   8   9  10  11  12   13  14  15  16   17  
#18  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40  41  42
lookup$DestCol =               c( NA, NA, "e", NA,  NA, "e","e","e","e","e","e","e", NA, "e","e","e", NA,               "e","e","g","i","e","e","e","f","g","e","e","e","e","e","e","e","e","e","e","e","e","e","e","e","e") 
lookup$DestSheet = as.numeric(c(  NA, NA,  2,  NA,  NA,  2,  2,  2,  2,  2,  3,  3,  NA,  3,  3,  3,  NA,  
 3,  3,  3,  3,  6,  3,  4,  4,  4,  3,  4,  6,  6,  6,  6,  6,  6,  6,  6,  6,  6,  6,  6,  6,  6))
lookup$DestRow =   as.numeric(c(  NA, NA,  6,  NA,  NA, 11, 12, 14, 17, 18,  5, 12,  NA, 13, 14, 15,  NA, 
16, 19, 19, 19, 61, 27, 12, 12, 12, 31, 16, 36, 38, 39, 41, 42, 43, 53, 46, 49, 50, 51, 54, 56, 60))

myLetters <- letters[1:26]  # used to match letter to number for destination column
lookup$DestCol <- match(lookup$DestCol, myLetters)

lookup$row_num <- seq.int(nrow(lookup)) 

```
  
Here's what we need to do to the data in order to send it to the correct cells in the Cool Farm Tool. This code is replicated in the 'for' loop in the subsequent code chunk.
```{r}
#rm(X)
#X = mgmt_data[1,]
#X <- t(X)
#class(X)
#X$row_num <- seq.int(nrow(X)) 
#Y <- cbind(X, seq.int(nrow(X)) ) 

#Z <- merge(Y, lookup, by = "row.names")
#Z <- Z[order(as.numeric(as.character(Z$V2))), ]
```


### CFT function
Now we write the function to send the management data to the CFT and extract the desired subset of the carbon emissions results.
```{r CFT-function}
run_CFT <- function(DF, subtract = "nothing"){
  # `if` statement to allow user to leave out any of the sources of emissions 
  # e.g. fertiliser use: 
  if(subtract == "herb fuel"){
    DF$applicationrate = 0
  }else if(subtract == "herb manuf"){
    DF$pesticide_applicns = 0
  }else if(subtract == "till"){
    DF[,which(colnames(CFTData)=="chisel_plough"):which(colnames(CFTData)=="tine_harrow")] = 0
  }else{
    DF = DF
  } 
  output = data.frame(fix.empty.names = FALSE)
  
  for (a in 1:nrow(DF)) {
    
    # `For` loop reads each row (field) and puts it into a writable format for CFT
    #X = mgmt_data[1,] # so X is the whole first row (then it will be the next row, etc)
    X = DF[a,] # the loop will read one row at a time.
    X <- t(X)  # transposes X to allow it to be merged with the lookup table
    
    
    FieldID = X["fieldID",]
    #FieldID = X[5,]
    print.default(a) # prints 1 for the first round of the loop, then 2 etc.
    
    # add row numbers so I can sort back into original order later
    X <- cbind(X, seq.int(nrow(X)) ) 

    X <- merge(X, lookup, by = "row.names") # Gives destination of each variable
    
    X <- X[!is.na(X$DestCol),] # Removes unnecessary data
    
    # Set column names
    X <- setNames(X, c("Variable","Value","OrigRow","DestSheet","DestCol","DestRow")) 
    
    # Sort back into original order
    X <- X[order(as.numeric(as.character(X$OrigRow))), ]
    
    # Remove column that's causing issues with the C stock change:
    #X <- X[X$Variable != "tillage_changes_time", ] 
    
    # If fertiliser type is to change within a scenario/ rotation, this will need 
    # to be incorporated into the input file and then written into 
    # the CFT in the loop below. In our case fertiliser is always the same 
    # in our scenarios, so we can just set it here:
    writeWorksheet(CFT,"N", 3, 19, 6, 
                   header = FALSE)  # Sets fertiliser nutrient     
    writeWorksheet(CFT,"Ammonium nitrate - 35% N", 3, 19, 5, 
                   header = FALSE) # Sets fertiliser type
    
    
     for(i in 1:nrow(X)){
      # `For` loop reads each row value and writes it to the destination cell 
      # as found from the lookup table
      
      b = X$Value[i]
      c = X$DestSheet[i]
      d = X$DestRow[i]
      e = X$DestCol[i]
      writeWorksheet(CFT, # object
                     b,   # data
                     c,   # sheet
                     d,   # startRow
                     e,   # startCol
                     header = FALSE)
      
      #print(paste("for row", i, "write CFT changes ", X$Variable[i], "to ", X$Value[i]))
      #print(paste("for row", i, "write CFT leads to Carbon change of: ", 
      #            readWorksheet(CFT, 9, 13, 8, 27, 8)[6, 1]))
      
    }
    
    #setForceFormulaRecalculation(CFT, sheet = c(2:18), TRUE)  # Forces excel to re-evaluate all formulae in sheets 2-18
    #Sys.sleep(1)
    # Read results from sheet 9, startRow = 13, startCol = 8, endRow = 27, 
    # endCol = 8 (returns a data.frame):
    results = readWorksheet(CFT,9,13,8,27,8) 
    results[is.na(results)] = 0
    #results <- rbind(results, c(sum(results$Per.hectare))) 
    results <- setNames(results, "Value")
    results <- data.frame(c(FieldID, results$Value[1], results$Value[2],
                            results$Value[4], results$Value[6], 
                            results$Value[10], results$Value[14]),
                          fix.empty.names = FALSE) # Extracts desired results
    output <- rbind(output,t(results))
    
  }
  output <- setNames(output, c("Field ID", 
                               "Fertiliser production (kgC02eq/ha)",
                               "Field N20 from fertiliser (kgC02eq/ha)", #Direct and indirect field N20 from fertiliser
                               "Pesticide (kgC02eq/ha)", 
                               "C stock changes (kgC02eq/ha)", 
                               "Field energy use (kgC02eq/ha)", 
                               "Total (kgC02eq/ha)")) # renames columns 
  return(output)
}
```

### Calculate emissions
```{r run-CFT}
#output_data <- run_CFT(mgmt_data[1:34,])
output_data <- run_CFT(mgmt_data)
write.csv(output_data, "../output/carbon_results.csv", row.names = FALSE)
#output_data
```
