==> scripts/07. Analysis.Rmd <==
---
title: "Analyse trade-offs"
author: "Alexa Varah"
date: "2024-05-21"
output: html_document
---

# 1. Intro
  
This script looks at what happens to farm gross profits, yield (in calories), and carbon emissions when farmers switch from a BAU crop rotation to strategies aimed at mitigating herbicide-resistance black-grass in England.  
  
We ignore the 'continuous winter wheat' scenario from the previous paper as we showed there that it's worst for controlling herbicide-resistant black-grass.  

Data `c_gp_yld` is already summarised across imputations.  


## Sources of emissions in CFT

**Pesticide emissions**  
Pesticide emissions = number of pesticide applications * embodied emissions in pesticides (kg COs-eq / dose / ha) * crop yield / ha  
This doesn't account for the fuel used when applying pesticides, only the emissions caused by pesticide production.  

**Direct and indirect field N2O**  
Direct and indirect field N2O = Background direct and indirect N2O + Fertiliser induced field emissions (FIE)  

Background direct and indirect N2O = N2O + NO (they don't include NH3 volatilisation + leaching in this equation)
Fertiliser induced field emissions = N to N2O conversion factor * (N2O + NO + NH3 volatilisation + leaching)
NH3 volatilisation = volatilisation emissions factor of 0.01 * 
                      sum of NH3 volatilisation from each fertiliser application
NH3 volatilisation from a fertiliser application = amount of N applied * 
      EXP(sum of NH3 volatilisation factors for climate, crop type, soil texture, SOM, soil CEC, drainage, pH + 
              NH3 from fertiliser production emissions + 
                NH3 from fertiliser application)
              
**Field energy use**  
This contains emissions from diesel used in tillage, herb spraying, fert spreading/spraying, other field ops (harvest etc - we haven't input data on these other field ops).

  
# 2. Load packages, functions, data
## 2.1. Load packages & functions
```{r}
# Load packages
rm(list = ls(all = TRUE)) # clear envt
#library(hrbrthemes) # for theme_ipsum() in plot
remove.packages("tidyverse")
install.packages(c("ggplot2", "dplyr", "tidyr", "readr", "purrr", "tibble", "stringr"))
install.packages("tidyverse")
library(tidyverse) # for east data manip and plotting
library(patchwork) # for panel plots
library(grateful) # for citing R packages
library(readxl) # for reading in Excel workbooks / spreadsheets
library(viridis) # for colour-blind-friendly palettes
library(forcats) # for easy re-coding of factors

# Load functions
source("functions/Count_complete_observations.R")
#source("functions/Annotate_plots.R")
```

## 2.2. Load data
```{r}
# Load data on profit, yield & carbon----
c_gp_yld <- read.csv("../output/trade_offs_tidy.csv", header = T) %>% 
  dplyr::mutate(
    scenario = factor(scenario, levels = c("BAU", "MIT")),
    region = factor(region, levels = c("north", "central", "east")),
    soiltype = factor(soiltype, levels = c("heavy", "medium","light")),
    initDR = factor(initDR, levels = c("LD-LR", "LD-HR", "HD-HR"))
    )

# Load management strategy data----
strat <- read_xlsx("../data/strategy_design_summary.xlsx",2) %>% 
  dplyr::mutate(
    scenario = as.factor(scenario),
    design_element = as.factor(design_element),
    initdenres = as.factor(initdenres),
    region = as.factor(region),
    soil_type = as.factor(soil_type)
  ) 

# Load tidied density simulation data, all 37 imputations----
# Sum high and very high density proportions (-> hvh) across 3 density-resistance levels.  
res_tidy <- readRDS(file = "../data/simulation_results_tidy.rds") %>%  
  dplyr::filter(scenario != "CWW") %>% 
  dplyr::mutate( 
    scenario = factor(scenario, levels = c("BAU", "MIT")),
    initden = factor(initden, levels = c("VD", "HD", "MD", "LD")),
    initres = factor(initres),
    region = factor(region, levels = c("north", "central", "east")),
    soiltype = factor(soiltype, levels = c("heavy", "medium", "light")),
    initdenres = factor(initdenres, # specify order of levels
                        levels = c("VD-HR", "HD-HR", "MD-HR", 
                                   "MD-LR", "LD-HR", "LD-LR")),
    initcondit = factor(initcondit, levels = c(
      "n_h_VD-HR", "n_m_VD-HR", "n_l_VD-HR", 
      "c_h_VD-HR", "c_m_VD-HR", "c_l_VD-HR",
      "e_h_VD-HR", "e_m_VD-HR", "e_l_VD-HR",
      
      "n_h_HD-HR", "n_m_HD-HR", "n_l_HD-HR",
      "c_h_HD-HR", "c_m_HD-HR", "c_l_HD-HR",
      "e_h_HD-HR", "e_m_HD-HR", "e_l_HD-HR",
      
      "n_h_MD-HR", "n_m_MD-HR", "n_l_MD-HR",
      "c_h_MD-HR", "c_m_MD-HR", "c_l_MD-HR",
      "e_h_MD-HR", "e_m_MD-HR", "e_l_MD-HR",
      
      "n_h_MD-LR", "n_m_MD-LR", "n_l_MD-LR",
      "c_h_MD-LR", "c_m_MD-LR", "c_l_MD-LR",
      "e_h_MD-LR", "e_m_MD-LR", "e_l_MD-LR",
      
      "n_h_LD-HR", "n_m_LD-HR", "n_l_LD-HR",
      "c_h_LD-HR", "c_m_LD-HR", "c_l_LD-HR",
      "e_h_LD-HR", "e_m_LD-HR", "e_l_LD-HR",
      
      "n_h_LD-LR", "n_m_LD-LR", "n_l_LD-LR",
      "c_h_LD-LR", "c_m_LD-LR", "c_l_LD-LR",
      "e_h_LD-LR", "e_m_LD-LR", "e_l_LD-LR")),
    strategy = factor(strategy)
  ) %>%
  ungroup() %>%
  # collapse init den res to three levels
  dplyr::mutate(
    initDR3 = forcats::fct_recode(initdenres, # collapse categories
                                  "LD-LR" = "MD-LR", # new, old
                                  "LD-HR" = "MD-HR", 
                                  "HD-HR" = "VD-HR"),
    initDR3 = factor(initDR3, levels = c("LD-LR", "LD-HR", "HD-HR"))
                ) %>% 
  ungroup() %>% 
  # sum the proportions of high and very high density black-grass:
  rowwise() %>% 
  dplyr::mutate(hvh = sum(c_across(h:v), na.rm = T)) %>%  # new category: h & vh
  ungroup() %>% 
  dplyr::rename(mds = mds.ms)
```

# 3. Calculate derived variables / summary stats

## 3.1. Management variables
Calculate total volume of glyphosate, total vol selective herbs, and total number of herb applications, in a rotation. 
```{r}
strat2 <- strat %>% 
  # calculate rotation totals
  filter(design_element %in% c("gly_vol_yr", "selherb_vol_yr", "herb_applic_n_yr")) %>%
  dplyr::group_by(scenario, design_element, initdenres, region, soil_type) %>% 
  summarise(value = sum(value, na.rm = TRUE)) %>% 
  ungroup() %>% 
  dplyr::mutate(
    # rename design elements
    design_element = forcats::fct_recode(design_element, 
                                         "gly_rot_vol" = "gly_vol_yr",
                                         "selherb_rot_vol" = "selherb_vol_yr",
                                         "herb_applic_n_rot" = "herb_applic_n_yr"),
    # add 'year' and 'scaled_value' columns so can bind to original data frame
    year = 'NA'
    ) %>% 
  dplyr::relocate(year, .before = value)

colnames(strat)
colnames(strat2)
# join the two dataframes
strat <- rbind(strat, strat2)

strat <- strat %>% 
  # Within each design element, scale values between 0 and 1:
  group_by(design_element) %>%
  dplyr::mutate(scaled_value = (value - min(value)) / (max(value) - min(value))) %>%
  ungroup() 

rm(strat2)
```

## 3.2. Summarise final weed density, gross profit, opp costs, yield & carbon across rotations 
For plots of the effect of initial density-resistance state *within regions and soil types* (i.e., panel plots, where panels are region and soil), range needs to show variability across years in the rotation.  

### 3.2.1. Carbon, gross profit, yield (rotation mean and differences between BAU and MIT)
Here, the mean is the rotation mean and the min and max values give range across years within a rotation.  
```{r, attr.source='.numberLines'}
trdoff_rot <- c_gp_yld %>% 
  dplyr::group_by(scenario, region, soiltype, initDR) %>% 
  # calculate min and max values (this grouping will give range across years within a rotation)----
  dplyr::mutate(
    gp_min = min(gp, na.rm = TRUE),
    gp_max = max(gp, na.rm = TRUE),
    oc_min = min(oc, na.rm = TRUE),
    oc_max = max(oc, na.rm = TRUE),
    ww_pc_min = min(ww_pc, na.rm = TRUE),
    ww_pc_max = max(ww_pc, na.rm = TRUE),
    calories_min = min(calories, na.rm = TRUE),
    calories_max = max(calories, na.rm = TRUE),
    calories_diff_min = min(calories_diff, na.rm = TRUE),
    calories_diff_max = max(calories_diff, na.rm = TRUE),
    yield_min = min(yield, na.rm = TRUE),
    yield_max = max(yield, na.rm = TRUE),
    land_use_min = min(land_use, na.rm = TRUE),
    land_use_max = max(land_use, na.rm = TRUE),
    land_use_diff_min = min(land_use_diff, na.rm = TRUE),
    land_use_diff_max = max(land_use_diff, na.rm = TRUE),
    c_total_min = min(c_total, na.rm = TRUE),
    c_total_max = max(c_total, na.rm = TRUE),
    c_emiss_diff_min = min(c_emiss_diff, na.rm = TRUE),
    c_emiss_diff_max = max(c_emiss_diff, na.rm = TRUE)
  ) %>% 
  # calculate mean values---- 
  dplyr::summarise(across(c(
    gp, gp_min, gp_max, # all min values are the same so the mean == min, ditto for max
    oc, oc_min, oc_max, 
    ww_pc, ww_pc_min, ww_pc_max, 
    calories, calories_min, calories_max,
    calories_diff, calories_diff_min, calories_diff_max,
    yield, yield_min, yield_max,
    land_use, land_use_min, land_use_max,
    land_use_diff, land_use_diff_min, land_use_diff_max,
    c_total, c_total_min, c_total_max,
    c_emiss_diff, c_emiss_diff_min, c_emiss_diff_max
    ),.fns = list(
      mean=~mean(.x, na.rm = TRUE),
      n=~nobs(.x)))) %>% 
  ungroup() %>% 
  # tidy up data frame----
  dplyr::select(-contains(c('min_n', 'max_n'))) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("min_mean", "min", .x)) %>% # remove '_mean' in 'minimum' column names
  dplyr::rename_at(vars(everything()), ~ sub("max_mean", "max", .x)) %>% # remove '_mean' in 'maximum' column names
  dplyr::rename(n = oc_n) %>% # create 1 column to hold 'n'
  dplyr::relocate(n, .after = initDR) %>% 
  dplyr::select(-contains('_n'))
```




### 3.2.2. Weed density ((a) rotation mean density and (b) proportion h & vh in year 6)
This gives mean for each level of initial density and resistance. The min and max values show the lowest and highest rotation mean % change, or year 6 mean % change, across regions and soil types.  
```{r, attr.source='.numberLines'}
density_rot <- res_tidy %>%
  dplyr::rename(initDR = initDR3) %>% # rename the column
  
  # Sum the proportions of high and very high density black-grass
  rowwise() %>% 
  dplyr::mutate(hvh = sum(c_across(h:v), na.rm = T)) %>%  # new category: h & vh
  ungroup() %>% 
  
  # Specify order of factor levels and arrange df
  dplyr::mutate( # specify factor levels
    scenario = factor(scenario, levels = c("BAU", "MIT")),
    region = factor(region, levels = c("north", "central", "east")),
    soiltype = factor(soiltype, levels = c("heavy", "medium","light")),
    initDR = factor(initDR, levels = c("LD-LR", "LD-HR", "HD-HR"))
    ) %>% 
  
  # Calculate mean across imputations for each year of each strategy
  group_by(scenario, initDR, region, soiltype, year) %>% 
  dplyr::mutate(
    mds_mean = mean(mds), # mean density state across imputations
    hvh_mean = mean(hvh) # mean proportion of field with high or very high density across imputations
  ) %>% 
  ungroup() %>% 
  distinct(scenario, initDR, region, soiltype, year, .keep_all = TRUE) %>% # retain only one unique row per strategy-year
  dplyr::select(scenario, initDR, region, soiltype, year, mds_mean, hvh_mean) %>% # pull out useful columns
  dplyr::rename(mds = mds_mean, hvh = hvh_mean) %>% # rename columns
  ungroup() %>% 
  dplyr::arrange(scenario, initDR, region, soiltype, year) %>% 
  
  # Calculate rotation mean density
  dplyr::group_by(scenario, initDR, region, soiltype) %>% 
  dplyr::mutate(mds_rot_mean = mean(mds)) %>% 
  ungroup() %>% 
  dplyr::filter(year == 6) %>% # retain just year 6
  
  # Calculate difference between MIT and BAU for rotation means and for yr6
  dplyr::group_by(initDR, region, soiltype) %>% 
  dplyr::mutate(
    mds_diff = mds_rot_mean - mds_rot_mean[scenario == "BAU"],
    hvhyr6_diff = hvh - hvh[scenario == "BAU"]
  ) %>% 
  ungroup() %>% 

  # Calculate percentage change
  dplyr::group_by(region, soiltype, initDR) %>% 
  dplyr::mutate(
    mds_pc = (mds_diff/mds_rot_mean[scenario == "BAU"])*100, # % change in rotation mean density
    hvhyr6_pc = (hvhyr6_diff/hvh[scenario == "BAU"])*100 # % change in proportion of hvh at year 6
  ) %>% 
  ungroup() %>% 
  
  # Calculate min and max values (this grouping will give range across region and soil types)
  dplyr::filter(scenario == "MIT") %>% 
  dplyr::group_by(initDR) %>%
  dplyr::mutate(
    # min & max % change in rotation mean density
    mds_pc_min = min(mds_pc, na.rm = TRUE), 
    mds_pc_max = max(mds_pc, na.rm = TRUE), 
    # min & max % change in proportion of hvh at year 6
    hvhyr6_pc_min = min(hvhyr6_pc, na.rm = TRUE), 
    hvhyr6_pc_max = max(hvhyr6_pc, na.rm = TRUE), 
  ) %>%
  ungroup() %>%
  dplyr::select(initDR, region, soiltype, mds_pc:hvhyr6_pc_max) %>% 
  
  # Calculate mean for each level of initDR
  dplyr::group_by(initDR) %>% 
  dplyr::mutate(
    mds_pc_mean = mean(mds_pc), # mean % change in rotation mean density
    hvhyr6_pc_mean = mean(hvhyr6_pc), # mean % change in rotation mean density
  ) %>% 
  ungroup() %>% 
  
  # Tidy the data frame
  distinct(initDR, .keep_all = TRUE) %>% # retain only one unique row per initial density-resistance state
  dplyr::select (initDR, mds_pc_mean, mds_pc_min, mds_pc_max, hvhyr6_pc_mean, hvhyr6_pc_min, hvhyr6_pc_max) # retain required columns
```


## 3.3. Emissions sources 

Sources of GHG emissions:  
c_fert_prod  
c_fert_appl  
c_pesticide     
c_field_energy_use  


*N.B.* There are no estimates from crop residues. In the CFT I specified 'left on field'; however, I didn't specify the amount of residue. The CFT says that a default amount will be calculated, but in fact this doesn't happen. If I wanted to re-run the analysis to include amounts for residues, I could calculate them using the equations provided in Scarlat 2010 (https://www-sciencedirect-com.ezproxy.nhm.ac.uk/science/article/pii/S0956053X10002436). Use the AHDB webside or maybe https://www.researchgate.net/publication/256793081_Straw_Use_and_Availability_for_Second_Generation_Biofuels_in_England for sugar beet or any other crops not in Scarlat 2010.

*N.B.* I didn't include harvesting field energy use. This is because all but one crop used there is harvested by combine harvester. Beet is harvested with a beet harvester. Beet harvesters and combine harvesters have similar fuel use (see https://www.mdpi.com/2077-0472/13/4/833 for combine harvester fuel use and https://www.eadt.co.uk/news/business/21765742.east-anglian-farmers-focus-fuel-efficiency/ or 'EVALUATION OF FOSSIL FUEL CONSUMPTION AND ENVIRONMENTAL IMPACTS IN SUGAR BEET PRODUCTION AS A BIOFUEL RAW MATERIAL' Fresenius Environmental Bulletin Volume 28 ± No. 4/2019 pages 2397-2406 for beet harvester fuel use). ALTHOUGH they do have different values in the Cool Farm Tool, so perhaps I should have included harvesting operations.....

Calculate differences due to each source.
```{r}
c_source <- c_gp_yld %>% 
  # find differences between BAU and MIT for each GHG source
  dplyr::group_by(region, soiltype, initDR, year) %>% 
  dplyr::mutate(
    # c_fert_prod
    c_fert_prod_diff = c_fert_prod - c_fert_prod[scenario == "BAU"],
    c_fert_prod_pc_change = (c_fert_prod_diff/c_fert_prod[scenario == "BAU"])*100,
    # c_fert_appl
    c_fert_appl_diff = c_fert_appl - c_fert_appl[scenario == "BAU"],
    c_fert_appl_pc_change = (c_fert_appl_diff/c_fert_appl[scenario == "BAU"])*100,
    # c_pesticide
    c_pesticide_diff = c_pesticide - c_pesticide[scenario == "BAU"],
    c_pesticide_pc_change = (c_pesticide_diff/c_pesticide[scenario == "BAU"])*100,
    # c_field_energy_use
    c_feu_diff = c_field_energy_use - c_field_energy_use[scenario == "BAU"],
    c_feu_pc_change = (c_feu_diff/c_field_energy_use[scenario == "BAU"])*100,
  ) %>% 
  dplyr::ungroup() %>% 
  dplyr::filter(scenario == 'MIT') %>% 
   
  # get min and max values (this grouping will give range across years)----
  dplyr::group_by(region, soiltype, initDR) %>% 
  dplyr::mutate(
    # c_fert_prod
    c_fert_prod_diff_min = min(c_fert_prod_diff, na.rm = TRUE),
    c_fert_prod_diff_max = max(c_fert_prod_diff, na.rm = TRUE),
    # c_fert_appl
    c_fert_appl_diff_min = min(c_fert_appl_diff, na.rm = TRUE),
    c_fert_appl_diff_max = max(c_fert_appl_diff, na.rm = TRUE),
    # c_pesticide
    c_pesticide_diff_min = min(c_pesticide_diff, na.rm = TRUE),
    c_pesticide_diff_max = max(c_pesticide_diff, na.rm = TRUE),
    # c_field_energy_use
    c_feu_diff_min = min(c_feu_diff, na.rm = TRUE),
    c_feu_diff_max = max(c_feu_diff, na.rm = TRUE)
  ) %>% 
  dplyr::ungroup() %>% 
  dplyr::select(ID:crop, c_fert_prod_diff:c_feu_diff_max) %>% 
  
  # calculate mean values across a rotation---- 
dplyr::group_by(scenario, region, soiltype, initDR) %>% 
  dplyr::summarise(across(c(
    # c_fert_prod
    c_fert_prod_diff, c_fert_prod_diff_min, c_fert_prod_diff_max,
    # c_fert_appl
    c_fert_appl_diff, c_fert_appl_diff_min, c_fert_appl_diff_max, 
    # c_pesticide
    c_pesticide_diff, c_pesticide_diff_min, c_pesticide_diff_max, 
    # c_field_energy_use
    c_feu_diff, c_feu_diff_min, c_feu_diff_max
    ),.fns = list(
      mean=~mean(.x, na.rm = TRUE),
      n=~nobs(.x)))) %>% 
  ungroup() %>% 
  
  # tidy up----
  dplyr::select(-contains(c('min_n', 'max_n'))) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("min_mean", "min", .x)) %>% # remove '_mean' in 'minimum' column names
  dplyr::rename_at(vars(everything()), ~ sub("max_mean", "max", .x)) %>% # remove '_mean' in 'maximum' column names
  dplyr::rename(n = c_fert_prod_diff_n) %>% # create 1 column to hold 'n'
  dplyr::relocate(n, .after = initDR) %>% 
  dplyr::select(-contains('_n')) %>% 
  dplyr::ungroup()

c_source
```




Convert from wide to long for plotting.
```{r}
mean_df <- c_source %>%
  dplyr::select(region, soiltype, initDR, contains('mean')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_mean", "", .x)) %>% # remove '_diff_mean' from column names
  dplyr::rename_at(vars(everything()), ~ sub("c_", "", .x)) %>% # remove 'c_' from column names
  pivot_longer(
    #cols = ends_with("mean"),
    cols = fert_prod:feu,
    names_to = "source_of_emissions",
    #names_suffix = "_mean",
    values_to = "rotation_mean",
    values_drop_na = TRUE
  )

min_df <- c_source %>%
  dplyr::select(region, soiltype, initDR, contains('min')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_min", "", .x)) %>% # remove '_diff_mean' from column names
  dplyr::rename_at(vars(everything()), ~ sub("c_", "", .x)) %>% # remove 'c_' from column names
  pivot_longer(
    cols = fert_prod:feu,
    names_to = "source_of_emissions",
    values_to = "rotation_min",
    values_drop_na = TRUE
  )

max_df <- c_source %>%
  dplyr::select(region, soiltype, initDR, contains('max')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_max", "", .x)) %>% # remove '_diff_mean' from column names
  dplyr::rename_at(vars(everything()), ~ sub("c_", "", .x)) %>% # remove 'c_' from column names
  pivot_longer(
    cols = fert_prod:feu,
    names_to = "source_of_emissions",
    values_to = "rotation_max",
    values_drop_na = TRUE
  )

# Join data frames
sources <- dplyr::left_join(mean_df, min_df, by = c("region", "soiltype", "initDR", "source_of_emissions")) %>% 
  dplyr::left_join(., max_df, by = c("region", "soiltype", "initDR", "source_of_emissions")) %>% 
  dplyr::mutate(across('source_of_emissions', ~ str_replace(., 'pesticide', 'herbicide'))) %>% 
  dplyr::mutate(across('source_of_emissions', ~ str_replace(., 'feu', 'field_energy_use'))) %>% 
  dplyr::mutate(across('source_of_emissions', ~ str_replace(., 'fert_appl', 'field_N20_from_fertiliser'))) %>% 
  dplyr::mutate(across('source_of_emissions', ~ str_replace(., 'fert_prod', 'fertiliser_production'))) 
  
# Tidy up
rm(mean_df, min_df, max_df, c_source)
```




# 4. Tables
## 4.1. Mean difference in C emissions per initDR 
### 4.1.1. Calculate difference for each year, take mean across rotation
(a) WRONG This gives min and max values across regions and soil types.  
```{r}
c_gp_yld %>%
  
  # 1. Calculate rotation mean gp, yield, landuse, GHG emiss
  dplyr::group_by(scenario, region, soiltype, initDR) %>% 
  dplyr::mutate(
    ghg_mean = mean(c_total)
  ) %>% 
  ungroup() %>% 
  dplyr::select(c(scenario:initDR, ghg_mean)) %>% 
  dplyr::distinct(scenario, region, soiltype, initDR, .keep_all = TRUE) %>% 
  
  # 2. Calculate difference between MIT and BAU
  dplyr::group_by(region, soiltype, initDR) %>% 
  dplyr::mutate(
    ghg_diff = ghg_mean - ghg_mean[scenario == "BAU"]
  ) %>% 
  ungroup() %>% 
  
  # 3. Calculate min and max values (this grouping will give range across region and soil types)
  dplyr::filter(scenario == "MIT") %>% 
  dplyr::group_by(initDR) %>%
  dplyr::mutate(
    ghg_diff_min = min(ghg_diff, na.rm = TRUE),
    ghg_diff_max = max(ghg_diff, na.rm = TRUE)
  ) %>%
  ungroup() %>%
  
  # 5. Calculate mean for each level of initDR
  dplyr::group_by(initDR) %>% 
  dplyr::mutate(
    ghg_diff_mean = mean(ghg_diff)
  ) %>% 
  
  #Select just useful columns and filter out duplicates
  dplyr::select(initDR, 
                ghg_diff_mean, ghg_diff_min, ghg_diff_max) %>% 
  dplyr::distinct(initDR, .keep_all = TRUE) %>% 
  arrange(initDR)
```

Maxima for LD-HR:
```{r}
c_gp_yld %>%
  filter(scenario == 'MIT', initDR == 'LD-HR') %>%
  dplyr::group_by(scenario, region, soiltype, initDR) %>%
  mutate(max_diff = max(c_emiss_diff, na.rm = TRUE)) %>%
  select(scenario, region, soiltype, initDR, max_diff) %>%
  dplyr::distinct(scenario, region, soiltype, initDR, .keep_all = TRUE)
```
The highest maximum difference for LD-HR across all regions (n=3) and soil types (n=3) is 1136.96 (this is from east, heavy).

Minima for LD-HR:
```{r}
c_gp_yld %>%
  filter(scenario == 'MIT', initDR == 'LD-HR') %>%
  dplyr::group_by(scenario, region, soiltype, initDR) %>%
  mutate(min_diff = min(c_emiss_diff, na.rm = TRUE)) %>%
  select(scenario, region, soiltype, initDR, min_diff) %>%
  dplyr::distinct(scenario, region, soiltype, initDR, .keep_all = TRUE)
```
The lowest maximum difference for LD-HR across all regions (n=3) and soil types (n=3) is -2790.15 (this is from east, heavy).


Emissions differ by soil type (due to different fertiliser use):
```{r}
c_gp_yld %>%
  filter(scenario == 'MIT', region == 'north', initDR == 'HD-HR') %>%
  select(scenario, region, soiltype, initDR, year, crop, c_emiss_diff) %>%
  print(n = nrow(.))
```

(b) Calculate annual mean using c_emiss_diff (which is calculated for each year separately).
```{r}
c_gp_yld %>%
  dplyr::filter(scenario == "MIT") %>% 
  
  # 1. Get rotation mean, plus annual min and max differences for each strategy
  dplyr::group_by(scenario, region, soiltype, initDR) %>% 
  dplyr::mutate(
    ghg_diff_annual_mean = mean(c_emiss_diff),
    ghg_diff_annual_min = min(c_emiss_diff), # this is the minimum difference of any year for a particular rotation
    ghg_diff_annual_max = max(c_emiss_diff)
  ) %>%
  ungroup() %>%
  select(scenario, region, soiltype, initDR, ghg_diff_annual_mean, ghg_diff_annual_min, ghg_diff_annual_max) %>%
  dplyr::distinct(scenario, region, soiltype, initDR, .keep_all = TRUE) %>% 
    
  # 2. Calculate mean, min, max for each level of initDR
  dplyr::group_by(initDR) %>% 
  dplyr::mutate(
    ghg_annualdiff_mean = mean(ghg_diff_annual_mean),
    # min and max of the rotation means
    ghg_annualdiff_min = min(ghg_diff_annual_min),
    ghg_annualdiff_max = max(ghg_diff_annual_max),
    # mean across all the min (or max) values of any year for any rotation in each level of initDR
    # in other words, the mean value across the lowest (or highest) annual differences from each rotation
    # or, the mean across the minimum differences from any year for each particular rotation within one initDR level
    ghg_annualdiff_min_mean = mean(ghg_diff_annual_min),
    ghg_annualdiff_max_mean = mean(ghg_diff_annual_max),
    # min and max difference from any rotation in each level of initDR
    overall_min = min(ghg_diff_annual_min),
    overall_max = max(ghg_diff_annual_max)
  ) %>% 
  
  # 4. Select just useful columns and filter out duplicates
  dplyr::select(initDR, 
                ghg_annualdiff_mean, ghg_annualdiff_min, ghg_annualdiff_max, 
                ghg_annualdiff_min_mean, ghg_annualdiff_max_mean,
                overall_min, overall_max) %>% 
  dplyr::distinct(initDR, .keep_all = TRUE) %>% 
  arrange(initDR)
```


### 4.1.2. Sum over rotation, calculate difference
Try doing it another way. Instead of working out the difference for each year of the rotation and then taking the mean difference, sum across years to give a rotation total and then find the difference.
```{r}
# c_gp_yld %>%
#   
#   # 1. Sum over years
#   # 1. Calculate rotation mean gp, yield, landuse, GHG emiss
#   dplyr::group_by(scenario, region, soiltype, initDR) %>% 
#   dplyr::mutate(
#     ghg_sum = mean(c_total)
#   ) %>% 
#   ungroup() %>% 
#   dplyr::select(c(scenario:initDR, ghg_mean)) %>% 
#   dplyr::distinct(scenario, region, soiltype, initDR, .keep_all = TRUE) %>% 
#   
#   # 2. Calculate difference between MIT and BAU
#   dplyr::group_by(region, soiltype, initDR) %>% 
#   dplyr::mutate(
#     ghg_diff = ghg_mean - ghg_mean[scenario == "BAU"]
#   ) %>% 
#   ungroup() %>% 
#   
#   # 3. Calculate min and max values (this grouping will give range across region and soil types)
#   dplyr::filter(scenario == "MIT") %>% 
#   dplyr::group_by(initDR) %>%
#   dplyr::mutate(
#     ghg_diff_min = min(ghg_diff, na.rm = TRUE),
#     ghg_diff_max = max(ghg_diff, na.rm = TRUE)
#   ) %>%
#   ungroup() %>%
#   
#   # 5. Calculate mean for each level of initDR
#   dplyr::group_by(initDR) %>% 
#   dplyr::mutate(
#     ghg_diff_mean = mean(ghg_diff)
#   ) %>% 
#   
#   #Select just useful columns and filter out duplicates
#   dplyr::select(initDR, 
#                 ghg_diff_mean, ghg_diff_min, ghg_diff_max) %>% 
#   dplyr::distinct(initDR, .keep_all = TRUE) %>% 
#   arrange(initDR)
# 


c_gp_yld %>%
  
  # 1. Get total for each strategy
  dplyr::group_by(scenario, region, soiltype, initDR) %>% 
  dplyr::mutate(
    ghg_rot_tot = sum(c_total)
  ) %>%
  ungroup() %>%
  select(scenario, region, soiltype, initDR, ghg_rot_tot) %>%
  dplyr::distinct(scenario, region, soiltype, initDR, .keep_all = TRUE) %>% 
  
  # 2. Calculate difference between MIT and BAU
  dplyr::group_by(region, soiltype, initDR) %>% 
  dplyr::mutate(
    ghg_rot_tot_diff = ghg_rot_tot - ghg_rot_tot[scenario == "BAU"]
  ) %>% 
  ungroup() %>% 
    
  # 3. Calculate mean, min, max for each level of initDR
  dplyr::group_by(initDR) %>% 
  dplyr::mutate(
    ghg_rot_tot_diff_mean = mean(ghg_rot_tot_diff),
    ghg_rot_tot_diff_min = min(ghg_rot_tot_diff),
    ghg_rot_tot_diff_max = max(ghg_rot_tot_diff)
  ) %>% 
  
  # 4. Select just useful columns and filter out duplicates
  dplyr::select(initDR, 
                ghg_rot_tot_diff_mean, ghg_rot_tot_diff_min, ghg_rot_tot_diff_max) %>% 
  dplyr::distinct(initDR, .keep_all = TRUE) %>% 
  arrange(initDR)
```



## 4.2. Summary table in paper: Percentage change in yr 6 weed density, gp, C, calories, landuse, for each level of initDR
### 4.2.1. Calculate % change from rotation means for 3 levels of initDR
Here, I calculate the mean gp, kcal, land use, and GHG emissions across the rotation, then calculate the difference between BAU and MIT for each rotation, then calculate the % change for the rotation. The min and max values show the range across regions and soil types.  
```{r}
pc_change_rotn <- c_gp_yld %>%
  
  # 1. Calculate rotation mean for gp, yield, landuse, GHG emiss
  dplyr::group_by(scenario, region, soiltype, initDR) %>% 
  dplyr::mutate(
    # gross profit mean
    gp_mean = mean(gp),
    cal_mean = mean(calories),
    lu_mean = mean(land_use),
    ghg_mean = mean(c_total)
  ) %>% 
  ungroup() %>% 
  dplyr::select(c(scenario:initDR, gp_mean:ghg_mean)) %>% 
  dplyr::distinct(scenario, region, soiltype, initDR, .keep_all = TRUE) %>% 
  
  # 2. Calculate difference between MIT and BAU
  dplyr::group_by(region, soiltype, initDR) %>% 
  dplyr::mutate(
    gp_diff = gp_mean - gp_mean[scenario == "BAU"],
    cal_diff = cal_mean - cal_mean[scenario == "BAU"],
    lu_diff = lu_mean - lu_mean[scenario == "BAU"],
    ghg_diff = ghg_mean - ghg_mean[scenario == "BAU"]
  ) %>% 
  ungroup() %>% 
  
  # 3. Calculate percentage change
  dplyr::group_by(region, soiltype, initDR) %>% 
  dplyr::mutate(
    
    # gross profit % change
    gp_pc = (gp_diff/gp_mean[scenario == "BAU"])*100, # opportunity cost is already the difference in gross profit 
    
    # calories % change
    kcal_pc = (cal_diff/cal_mean[scenario == "BAU"])*100,    
    
    # land use % change
    landuse_pc = (lu_diff/lu_mean[scenario == "BAU"])*100,  
    
    # carbon emissions % change
    ghg_pc = (ghg_diff/ghg_mean[scenario == "BAU"])*100
    
  ) %>% 
  ungroup() %>% 
  
  # 4. Calculate min and max values (this grouping will give range across region and soil types)
  dplyr::filter(scenario == "MIT") %>% 
  dplyr::group_by(initDR) %>%
  dplyr::mutate(

    # gross profit difference min & max
    gp_pc_min = min(gp_pc, na.rm = TRUE),
    gp_pc_max = max(gp_pc, na.rm = TRUE),

    # calories difference min & max
    kcal_pc_min = min(kcal_pc, na.rm = TRUE),
    kcal_pc_max = max(kcal_pc, na.rm = TRUE),

    # land use difference min & max
    landuse_pc_min = min(landuse_pc, na.rm = TRUE),
    landuse_pc_max = max(landuse_pc, na.rm = TRUE),

    # GHG difference min & max
    ghg_pc_min = min(ghg_pc, na.rm = TRUE),
    ghg_pc_max = max(ghg_pc, na.rm = TRUE)

  ) %>%
  ungroup() %>%
  
  # 5. Calculate mean for each level of initDR
  dplyr::group_by(initDR) %>% 
  dplyr::mutate(
    
    # gross profit rotation mean % change
    gp_pc_mean = mean(gp_pc),

    # calories rotation mean % change
    kcal_pc_mean = mean(kcal_pc),
    
    # land use rotation mean % change
    lu_pc_mean = mean(landuse_pc),
    
    # carbon emissions rotation mean % change
    ghg_pc_mean = mean(ghg_pc)
    
  ) %>% 
  
  #Select just useful columns and filter out duplicates
  #dplyr::select(initDR, contains("pc_")) %>% 
  dplyr::select(initDR, 
                gp_pc_mean, gp_pc_min, gp_pc_max,
                kcal_pc_mean, kcal_pc_min, kcal_pc_max,
                lu_pc_mean, landuse_pc_min, landuse_pc_max,
                ghg_pc_mean, ghg_pc_min, ghg_pc_max) %>% 
  dplyr::distinct(initDR, .keep_all = TRUE) %>% 
  arrange(initDR)

pc_change_rotn
```

This approach (in which I calculated % change from rotation means) is the same as the approach used in the previous paper. This approach also makes more sense than the one below as here, the calories produced per m2 match up better with the landuse. If you produce more calories per m2 in MIT, then you should require less land to produce 1000 kcal. This is reflected here: where the % change is +ve for calories, it's -ve for landuse. Using the first approach, this relationship isn't there.

### 4.2.2. Add in pc change for weed density.
```{r}
pc_change_all <- pc_change_rotn %>% 
  dplyr::left_join(density_rot, by = "initDR") %>% 
  dplyr::select(-c(mds_pc_mean, mds_pc_min, mds_pc_max)) %>% 
  dplyr::relocate(hvhyr6_pc_mean:hvhyr6_pc_max, .after = initDR) %>% 
  dplyr::rename(
    bg_density_pc_mean = hvhyr6_pc_mean,
    bg_density_pc_min = hvhyr6_pc_min,
    bg_density_pc_max = hvhyr6_pc_max
  )

# Tidy up
rm(pc_change_rotn, density_rot)

pc_change_all
```

### 4.2.3. Save output as .csv
```{r}
write.csv(pc_change_all, "../output/2025-04-17_summarytable_percent_change.csv", row.names = FALSE)
```




# 5. Plots - main text
## 5.1. Diversification - histogram

Plot scaled rotation values of crop diversity, stale seedbeds, herbicide volumes:
```{r}
p_mgmt <- strat %>%
  dplyr::mutate(initdenres = fct_recode(initdenres, "BAU" = "NA")) %>% 
  dplyr::mutate(initdenres = factor(initdenres, levels=c("BAU", "LD-LR", "LD-HR", "HD-HR"))) %>%
  dplyr::mutate(ID = paste(scenario, initdenres)) %>% 
  dplyr:: group_by(ID) %>% 
  dplyr::filter(design_element %in% c("spring_crops_tot_n_rot", 
                                      "diff_crops_tot_n_rot", 
                                      "ssb_tot_n_rot", 
                                      "gly_rot_vol", 
                                      "selherb_rot_vol" )) %>% #"aut_gly_n_yr", 
  dplyr::mutate(design_element = factor(design_element, 
                                        levels=c("spring_crops_tot_n_rot", 
                                                 "diff_crops_tot_n_rot", 
                                                 "ssb_tot_n_rot",  
                                                 "gly_rot_vol", 
                                                 "selherb_rot_vol"))) %>%
  ggplot(aes(fill=design_element, y=scaled_value, x=initdenres)) +
    geom_boxplot(position=position_dodge2(preserve = "total"), outlier.colour="darkgrey", alpha=0.5) + #, outlier.colour="transparent",
    geom_point(color="lightgrey", size=1.8, width=0.1, position=position_jitterdodge(0.3)) + #, alpha=0.1
    #scale_fill_manual(values=c("#332288", "#44AA99", "#88CCEE", "#DDCC77", "#CC6677"),  
    scale_fill_manual(values=c("#44AA99", "#88CCEE", "#DDCC77", "#CC6677", "#AA4499"), #c("MIT" = "#40B0A6", "BAU" = "#E1BE6A"))
                      labels = c("Number of spring crops",
                                  "Number of different crops",
                                  "Number of stale seed beds",
                                  "Glyphosate volume",
                                  "Selective herbicide volume")) +
    # scale_fill_viridis(discrete=T, name="",
    #                    labels = c("Number of spring crops",
    #                               "Number of different crops",
    #                               "Number of stale seed beds",
    #                               "Glyphosate volume",
    #                               "Selective herbicide volume")
    #                    ) +
    guides(color = FALSE) + # suppresses the colour legend
    theme_classic() +
    theme(
        axis.title=element_text(size=26),
        axis.text.x=element_text(size=24), 
        axis.text.y=element_text(size=24),
        # axis.text.x=element_text(size=rel(2.2)), 
        # axis.text.y=element_text(size=rel(2.2)),
        #legend.title=element_text(size=22),
        legend.text=element_text(size=24),
        legend.title=element_blank()
        ) + 
    xlab("Scenario and initial density-resistance state") +
    ylab("Scaled value") +
    scale_x_discrete(labels=c("LD-LR" = "MIT\nLD-LR", 
                              "LD-HR" = "MIT\nLD-HR",
                              "HD-HR" = "MIT\nHD-HR")) +
    # These lines are thick enough and wide enough so none of the black line behind is visible.
    geom_segment(aes(x = 0.632, y = 0, xend = 0.77, yend = 0), colour = "#44AA99", size = 1.05) + # #332288
    geom_segment(aes(x = 0.783, y = 0, xend = 0.9175, yend = 0), colour = "#88CCEE", size = 1.05) + # #44AA99
    geom_segment(aes(x = 0.9325, y = 0, xend = 1.068, yend = 0), colour = "#DDCC77", size = 1.05) + # #88CCEE
    geom_segment(aes(x = 1.082, y = 0.0532, xend = 1.2168, yend = 0.0532), colour = "#CC6677", size = 1.05) + # #DDCC77
    geom_segment(aes(x = 1.233, y = 0.85865, xend = 1.3675, yend = 0.85865), colour = "#AA4499", size = 1.05)# #CC6677

p_mgmt

    # annotate("text", x=0.7, y=0.13, label="Spring crops", size = 3.2, angle=90, colour = "red") +
    # annotate("text", x=0.85, y=0.145, label="Different crops", size = 3.2, angle=90) +
    # annotate("text", x=1, y=0.155, label="Stale seed beds", size = 3.2, angle=90) +
    # annotate("text", x=1.15, y=0.23, label="Glyphosate volume", size = 3.2, angle=90) +    
    # annotate("text", x=1.29, y=0.66, label="Selective herb. volume", size = 3.2, angle=90) +
    # These lines all fit onto the existing black lines so that you can see a black border round the coloured line:
    # geom_segment(aes(x = 0.634, y = 0, xend = 0.7655, yend = 0), colour = "#332288") +
    # geom_segment(aes(x = 0.785, y = 0, xend = 0.916, yend = 0), colour = "#44AA99") +
    # geom_segment(aes(x = 0.935, y = 0, xend = 1.066, yend = 0), colour = "#88CCEE") +
    # geom_segment(aes(x = 1.084, y = 0.0534, xend = 1.2165, yend = 0.0532), colour = "#DDCC77") +
    # geom_segment(aes(x = 1.23465, y = 0.85865, xend = 1.3655, yend = 0.85865), colour = "#CC6677") 
```


Save plot
```{r}
# png("../figures/management_histogram.png",
#     height = 15, width = 18, units = "in", res = 300)
# print(p_mgmt)
# dev.off()

pdf("../figures/management_histogram.pdf", height = 15, width = 18)
print(p_mgmt)
dev.off()
```




## 5.2. Impact on weed density - violin plot

```{r}
# Plot LD-LR----
p_LDLR <- ggplot(res_tidy %>% filter(initDR3 == "LD-LR", year == 6), 
             aes(x = factor(scenario), y = hvh, fill = scenario)) +  
  geom_violin(alpha = 0.5) +  # Semi-transparent violin plots
  geom_jitter(width = 0.1, height = 0, color = "lightgrey", size = 0.5, alpha = 0.5) +  # Tiny, light grey, transparent points
  labs(x = "Scenario", y = "Proportion of field with high or very \nhigh density black-grass at year 6") +  # Custom axis titles
  scale_fill_manual(values = c("MIT" = "#44AA99", "BAU" = "#DDCC77")) +  # Custom colors for MIT and BAU (previously used: #40B0A6 and #E1BE6A)
  theme_minimal() +  # Base minimal theme for a clean look
  theme(
    panel.background = element_rect(fill = "white", color = NA),  # White background
    panel.grid = element_blank(),  # Remove gridlines
    axis.line = element_line(color = "black"),  # Black axis lines for both x and y
    axis.ticks = element_line(color = "black"),  # Black axis ticks
    axis.title=element_text(size=34),
    axis.text.x=element_text(size=30), 
    axis.text.y=element_text(size=30),
    # axis.text.x=element_text(size=rel(2.2)), 
    # axis.text.y=element_text(size=rel(2.2)),
    plot.title = element_text(size = 30, face = "bold"),
    legend.position = "none"  # Remove legend
  ) + 
  ggtitle("LD-LR") +
  ylim(0, 1)


# Plot LD-HR----
p_LDHR <- ggplot(res_tidy %>% filter(initDR3 == "LD-HR", year == 6), 
             aes(x = factor(scenario), y = hvh, fill = scenario)) +  
  geom_violin(alpha = 0.5) +  # Semi-transparent violin plots
  geom_jitter(width = 0.1, height = 0, color = "lightgrey", size = 0.5, alpha = 0.5) +  # Tiny, light grey, transparent points
  labs(x = "Scenario", y = "Proportion of field with high or \nvery high density black-grass at year 6") +  # Custom axis titles
  scale_fill_manual(values = c("MIT" = "#44AA99", "BAU" = "#DDCC77")) +  # Custom colors for MIT and BAU
  theme_minimal() +  # Base minimal theme for a clean look
  theme(
    panel.background = element_rect(fill = "white", color = NA),  # White background
    panel.grid = element_blank(),  # Remove gridlines
    axis.line = element_line(color = "black"),  # Black axis lines for both x and y
    axis.ticks = element_line(color = "black"),  # Black axis ticks
    axis.title=element_text(size=34),
    axis.text.x=element_text(size=30), 
    axis.title.y = element_blank(),
    axis.text.y=element_blank(),
    #axis.text.y=element_text(size=30),
    # axis.text.x=element_text(size=rel(2.2)), 
    # axis.text.y=element_text(size=rel(2.2)),
    plot.title = element_text(size = 30, face = "bold"),
    legend.position = "none"  # Remove legend
  ) + 
  ggtitle("LD-HR") 

# Plot HD-HR ----
p_HDHR <- ggplot(res_tidy %>% filter(initDR3 == "HD-HR", year == 6), 
             aes(x = factor(scenario), y = hvh, fill = scenario)) +  
  geom_violin(alpha = 0.5) +  # Semi-transparent violin plots
  geom_jitter(width = 0.1, height = 0, color = "lightgrey", size = 0.5, alpha = 0.5) +  # Tiny, light grey, transparent points
  labs(x = "Scenario", y = "Proportion of field with high or \nvery high density black-grass at year 6") +  # Custom axis titles
  scale_fill_manual(values = c("MIT" = "#44AA99", "BAU" = "#DDCC77")) +  # Custom colors for MIT and BAU
  theme_minimal() +  # Base minimal theme for a clean look
  theme(
    panel.background = element_rect(fill = "white", color = NA),  # White background
    panel.grid = element_blank(),  # Remove gridlines
    axis.line = element_line(color = "black"),  # Black axis lines for both x and y
    axis.ticks = element_line(color = "black"),  # Black axis ticks
    axis.title=element_text(size=34),
    axis.text.x=element_text(size=30), 
    axis.title.y = element_blank(),
    axis.text.y=element_blank(),
    #axis.text.y=element_text(size=30),
    # axis.text.x=element_text(size=rel(2.2)), 
    # axis.text.y=element_text(size=rel(2.2)),
    plot.title = element_text(size = 30, face = "bold"),
    legend.position = "none"  # Remove legend
  ) + 
  ggtitle("HD-HR") 
```


Make panel plot.
```{r}
bg_density_yr6 <- (p_LDLR + p_LDHR + p_HDHR) + plot_layout(axis_titles = "collect") 
bg_density_yr6
```


Save plot.
```{r}
# png("../figures/bg_density_yr6.png",
#     height = 6, width = 18, units = "in", res = 300)
# print(bg_density_yr6)
# dev.off()

pdf("../figures/bg_density_yr6.pdf", height = 10, width = 25)
print(bg_density_yr6)
dev.off()
```


How many points are in each violin plot?
```{r}
nrow(res_tidy %>% filter(scenario == "MIT", initDR3 == "HD-HR", year == 6))
```


## 5.3. Profit, calories, GHG emissions - panel plots

### 5.3.1. Direction of signs - what wins?
Difference in *gross profit: +ve value wins*  
Difference in gross profit (opp cost) = money lost by switching = difference in gross profit MIT-BAU.  
So if  
MIT = 10 <- better  
BAU = 5  
MIT - BAU = 5 = +ve value means MIT wins  
  
Difference in *carbon emissions: -ve value wins*  
Difference in carbon emissions MIT-BAU  
So if  
MIT = 5 <- better  
BAU = 8  
MIT - BAU = -3 = -ve value means MIT wins  
  
Difference in *calories: +ve value wins*  
Difference in kcal produced per m2 = MIT-BAU  
So if  
MIT = 6  <- better  
BAU = 3  
MIT - BAU = 3 = +ve value means MIT wins  
  
Difference in *land use: -ve value wins*    
Difference in amount of land used to produce 1000 kcal = MIT-BAU
So if  
MIT = 3  <- better  
BAU = 6  
MIT - BAU = -3 = -ve value means MIT wins
  
Could switch the sign on opportunity costs (would then be BAU-MIT) so that for both C emissions and gross profit, values below 0 mean MIT wins.  

  
  
### 5.3.2. With region on top panels

Plot.
```{r, echo=FALSE}
# Specify panel labels
region.labs <- c("Northern England", "Central England", "Eastern England") # new
names(region.labs) <- c("north", "central", "east") # original

soiltype.labs <- c("Heavy soil", "Medium soil", "Light soil") # new
names(soiltype.labs) <- c("heavy", "medium", "light") # original

# Plot
## opp costs ----
a_oc_1col <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
               aes(initDR, oc_mean, colour = oc_mean < 0)) + 
  scale_colour_manual(values = c("#44AA99", "#CC6677"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ region,
             labeller = labeller(soiltype = soiltype.labs, region = region.labs),
             scales="free_y"
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Initial density and resistance of Black-grass population", 
       y = "Mean annual difference in gross profit (£/ha)") +
  geom_pointrange(aes(ymin = oc_min, ymax = oc_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        #axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none",
        tag.text = element_text(size=22)
        ) 

## yield ----
b_cal_1col <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
               aes(initDR, calories_diff_mean, colour = calories_diff_mean < 0)) + 
  scale_colour_manual(values = c("#44AA99", "#CC6677"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ region,
             labeller = labeller(soiltype = soiltype.labs, region = region.labs),
             scales="free_y"
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Initial density and resistance of Black-grass population", 
       y = "Mean annual difference in calories produced (kcal/ha)") +
  geom_pointrange(aes(ymin = calories_diff_min, ymax = calories_diff_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        #axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none"
        )

## GHG emissions ----
c_ghg_1col <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
               aes(initDR, c_emiss_diff_mean, colour = c_emiss_diff_mean > 0)) + 
    scale_colour_manual(values = c("#44AA99", "#CC6677"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ region,
             labeller = labeller(soiltype = soiltype.labs, 
                                 region = region.labs)) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Initial density and resistance of Black-grass population", 
       y = expression("Mean annual difference in GHG emissions (kgCO"["2"]~"eq/ha)"),
             scales="free_y"
             ) +
  geom_pointrange(aes(ymin = c_emiss_diff_min, ymax = c_emiss_diff_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "bottom"
        )

# Combine plots ----
facetpanel_1col <- a_oc_1col + b_cal_1col + c_ghg_1col + plot_layout(ncol = 1) +
  plot_annotation(tag_levels = 'a', tag_prefix = '(',  tag_suffix = ')') & 
  theme(plot.tag = element_text(size = 22, face="bold"))
```




Save.
```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_all_outcomes_1col.png",
#     height = 30, width = 10, units = "in", res = 300)
# print(facetpanel_1col)
# dev.off()

pdf("../figures/all_outcomes_1col.pdf", height = 30, width = 10)
print(facetpanel_1col)
dev.off()
```





### 5.3.3. With initial conditions on top panels

Plot.
```{r, echo=FALSE}
# Specify panel labels
denres.labs <- c("LD-LR", "LD-HR", "HD-HR") # new
names(denres.labs) <- c("LD-LR", "LD-HR", "HD-HR") # original

soiltype.labs <- c("Heavy soil", "Medium soil", "Light soil") # new
names(soiltype.labs) <- c("heavy", "medium", "light") # original

# Plot
## opp costs ----
a_oc_alt <- ggplot(trdoff_rot %>% 
                     dplyr::filter(scenario == "MIT") %>% 
                             dplyr::mutate(region = fct_recode(region, N = "north", C = "central", E = "east")), 
               aes(region, oc_mean, colour = oc_mean < 0)) + 
  scale_colour_manual(values = c("#44AA99", "#CC6677"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs),
             scales="free_y"
             ) +
  #scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "English region", 
       y = "Mean annual difference in gross profit (£/ha)") +
  geom_pointrange(aes(ymin = oc_min, ymax = oc_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        #axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none",
        tag.text = element_text(size=22)
        ) 

## yield ----
b_cal_alt <- ggplot(trdoff_rot %>% 
                     dplyr::filter(scenario == "MIT") %>% 
                             dplyr::mutate(region = fct_recode(region, N = "north", C = "central", E = "east")), 
               aes(region, calories_diff_mean, colour = calories_diff_mean < 0)) + 
  scale_colour_manual(values = c("#44AA99", "#CC6677"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs),
             scales="free_y"
             ) +
  #scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "English region", 
       y = "Mean annual difference in calories produced (kcal/ha)") +
  geom_pointrange(aes(ymin = calories_diff_min, ymax = calories_diff_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        #axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none"
        )

## GHG emissions ----
c_ghg_alt <- ggplot(trdoff_rot %>% 
                     dplyr::filter(scenario == "MIT") %>% 
                             dplyr::mutate(region = fct_recode(region, N = "north", C = "central", E = "east")), 
               aes(region, c_emiss_diff_mean, colour = c_emiss_diff_mean > 0)) + 
    scale_colour_manual(values = c("#44AA99", "#CC6677"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)) +
  #scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "English region", 
       y = expression("Mean annual difference in GHG emissions (kgCO"["2"]~"eq/ha)"),
             scales="free_y"
             ) +
  geom_pointrange(aes(ymin = c_emiss_diff_min, ymax = c_emiss_diff_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "bottom"
        )

# Combine plots ----
facetpanel_denres_top <- a_oc_alt + b_cal_alt + c_ghg_alt + plot_layout(ncol = 1) +
  plot_annotation(tag_levels = 'a', tag_prefix = '(',  tag_suffix = ')') & 
  theme(plot.tag = element_text(size = 22, face="bold"))
```




Save.
```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_all_outcomes_1col.png",
#     height = 30, width = 10, units = "in", res = 300)
# print(facetpanel_1col)
# dev.off()

pdf("../figures/all_outcomes_1col_denres-top.pdf", height = 30, width = 10)
print(facetpanel_denres_top)
dev.off()
```




# 6. Plots - Supplementary Info
## 6.1. Change in GHG emissions by source

Specify panel labels.
```{r, specify-panel-labels, echo=FALSE}
denres.labs <- c("LD-LR", "LD-HR", "HD-HR") # new
names(denres.labs) <- c("LD-LR", "LD-HR", "HD-HR") # original

soiltype.labs <- c("Heavy soil", "Medium soil", "Light soil") # new
names(soiltype.labs) <- c("heavy", "medium", "light") # original
```

Make plots & combine.
-ve value means MIT wins (green = MIT wins, orange = MIT loses) 
```{r}
# north ----
sources_n_1col <- ggplot(sources %>% 
                           dplyr::filter(region == "north") %>% 
                           mutate(source_of_emissions = factor(source_of_emissions, levels = c("fertiliser_production", 
                                                                                               "field_N20_from_fertiliser", 
                                                                                               "field_energy_use",
                                                                                               "herbicide"))),
               aes(source_of_emissions, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#CC6677"), labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR, 
             labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs)#,
             #scales="free_y"
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of GHG emissions", 
       y = expression("Difference in GHG emissions (kgCO"[2]~"eq / ha)")
       #y = expression(atop("Mean annual difference in",
      #                    paste("GHG emissions (kgCO"["2"]~"eq / ha)"))) 
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none",
        tag.text = element_text(size=22)
        )


# central ----
sources_c_1col <- ggplot(sources %>% 
                           dplyr::filter(region == "central")%>% 
                           mutate(source_of_emissions = factor(source_of_emissions, levels = c("fertiliser_production", 
                                                                                               "field_N20_from_fertiliser", 
                                                                                               "field_energy_use",
                                                                                               "herbicide"))), 
               aes(source_of_emissions, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#CC6677"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of GHG emissions", 
       y = expression("Difference in GHG emissions (kgCO"[2]~"eq / ha)")
       #y = expression(atop("Mean annual difference in",
      #                    paste("GHG emissions (kgCO"["2"]~"eq / ha)"))) 
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none"
        ) 


# east ----
sources_e_1col <- ggplot(sources %>% 
                           dplyr::filter(region == "east")%>% 
                           mutate(source_of_emissions = factor(source_of_emissions, levels = c("fertiliser_production", 
                                                                                               "field_N20_from_fertiliser", 
                                                                                               "field_energy_use",
                                                                                               "herbicide"))), 
               aes(source_of_emissions, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#CC6677"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of GHG emissions", 
       y = expression("Difference in GHG emissions (kgCO"[2]~"eq / ha)")
       #y = expression(atop("Mean annual difference in",
      #                    paste("GHG emissions (kgCO"["2"]~"eq / ha)"))) 
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "bottom"
        ) 



# combine
sources_1col <- sources_n_1col + sources_c_1col + sources_e_1col + plot_layout(ncol = 1) +
  plot_annotation(
    #tag_levels = 'a', 
    tag_levels = list(c("(a) North", "(b) Central" , "(c) East")),
    tag_prefix = '',  
    tag_suffix = ''
    ) & 
  theme(plot.tag = element_text(size = 22, face="bold"))
```



Save plot.
```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_all_outcomes_1col.png",
#     height = 30, width = 10, units = "in", res = 300)
# print(facetpanel_1col)
# dev.off()

pdf("../figures/emissions_sources_1col.pdf", height = 30, width = 10)
print(sources_1col)
dev.off()
```


## 6.2. Difference in profit by cost component
### 6.2.1. Calculate the differences between BAU and MIT for required variables  
  
Herbicide cost (£/ha) = ('bgherb' application rate (l/ha) × bgherb price (£/l)) + ('glyphosate' application rate (l/ha) × (glyphosate price (£/l))  
Fertiliser cost (£/ha) = (N fertiliser rate (kg/ha) × N Fertiliser price (£/kg)) + (P fertiliser rate (kg/ha) × P fertiliser price (£/kg)) + (K fertiliser rate (kg/ha) × K fertiliser price (£/kg))  
Seed cost (£/ha) = Seed rate (kg/ha) × Seed price (£/kg)  
Sundry cost (£/ha) is primarily the cost of other chemicals and other miscellaneous costs.  
Variable cost (£/ha) = Fertiliser cost + Seed cost + Herbicide cost + Sundry cost  
Fuel cost (£/ha) = Machine work rate (h/ha) × Fuel consumption (l/h) × Fuel price (£/l)  
Labour cost (£/ha) = Work rate (h/ha) × Hourly labour wage (£/h)  
Operation cost (£/ha) = Fuel cost + Labour cost  
  
Use costs for herbicide, fertiliser, operations, seed, sundry. Also costs due to yield loss.  
```{r}
#rm(cost_differences)
cost_differences <- c_gp_yld %>% 

# 1. Calculate difference between MIT and BAU
  dplyr::group_by(initDR, region, soiltype, year) %>% 
  dplyr::mutate(
    output_diff = output - output[scenario == "BAU"], 
    cyl_diff = cyl - cyl[scenario == "BAU"], 
    herb_diff = herbicide_cost - herbicide_cost[scenario == "BAU"],
    fert_diff = fertiliser_cost - fertiliser_cost[scenario == "BAU"],
    ops_diff = operations_cost - operations_cost[scenario == "BAU"],
    seed_diff = seed_cost - seed_cost[scenario == "BAU"],
    sundry_diff = sundry_cost - sundry_cost[scenario == "BAU"]
  ) %>% 
  ungroup() %>% 

# 2. Get rotation min and max
  dplyr::filter(scenario == "MIT") %>% 
  dplyr::group_by(initDR, region, soiltype) %>% # this grouping will give range across years
  dplyr::mutate(
    # min & max difference in output
    output_diff_min = min(output_diff, na.rm = TRUE), 
    output_diff_max = max(output_diff, na.rm = TRUE), 
    # min & max difference in cost due to yield losses
    cyl_diff_min = min(cyl_diff, na.rm = TRUE), 
    cyl_diff_max = max(cyl_diff, na.rm = TRUE), 
    # min & max difference in cost due to herbicides
    herb_diff_min = min(herb_diff, na.rm = TRUE), 
    herb_diff_max = max(herb_diff, na.rm = TRUE), 
    # min & max difference in cost due to fertilisers
    fert_diff_min = min(fert_diff, na.rm = TRUE), 
    fert_diff_max = max(fert_diff, na.rm = TRUE), 
    # min & max difference in cost due to fuel, labour, operations
    ops_diff_min = min(ops_diff, na.rm = TRUE), 
    ops_diff_max = max(ops_diff, na.rm = TRUE), 
    # min & max difference in cost due to fuel, labour, operations
    seed_diff_min = min(seed_diff, na.rm = TRUE), 
    seed_diff_max = max(seed_diff, na.rm = TRUE), 
    # min & max difference in cost due to sundries and variables
    sundry_diff_min = min(sundry_diff, na.rm = TRUE), 
    sundry_diff_max = max(sundry_diff, na.rm = TRUE)
  ) %>%
  ungroup() %>%

# 3. Calculate rotation mean
  dplyr::group_by(initDR, region, soiltype) %>% 
  dplyr::mutate(
    output_diff_mean = mean(output_diff),
    cyl_diff_mean = mean(cyl_diff), 
    herb_diff_mean = mean(herb_diff), 
    fert_diff_mean = mean(fert_diff), 
    ops_diff_mean = mean(ops_diff), 
    seed_diff_mean = mean(seed_diff),
    sundry_diff_mean = mean(sundry_diff)
  ) %>% 
  ungroup() %>% 
  
# 4. Retain just useful rows/columns & tidy
  distinct(initDR, region, soiltype, .keep_all = TRUE) %>%
  dplyr::select (initDR, region, soiltype, 
                 output_diff_mean, output_diff_min, output_diff_max,
                 cyl_diff_mean, cyl_diff_min, cyl_diff_max,
                 herb_diff_mean, herb_diff_min, herb_diff_max,
                 fert_diff_mean, fert_diff_min, fert_diff_max,
                 ops_diff_mean, ops_diff_min, ops_diff_max,
                 seed_diff_mean, seed_diff_min, seed_diff_max,
                 sundry_diff_mean, sundry_diff_min, sundry_diff_max
                 )
```


### 6.2.2. Convert from wide to long for plotting
```{r}
mean_cost_diff_df <- cost_differences %>%
  dplyr::select(region, soiltype, initDR, contains('mean')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_mean", "", .x)) %>% # remove '_diff_mean' from column names
  pivot_longer(
    cols = output:sundry,
    names_to = "source_of_cost_difference",
    values_to = "rotation_mean",
    values_drop_na = TRUE
  )

min_cost_diff_df <- cost_differences %>%
  dplyr::select(region, soiltype, initDR, contains('min')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_min", "", .x)) %>% # remove '_diff_min' from column names
  pivot_longer(
    cols = output:sundry,
    names_to = "source_of_cost_difference",
    values_to = "rotation_min",
    values_drop_na = TRUE
  )

max_cost_diff_df <- cost_differences %>%
  dplyr::select(region, soiltype, initDR, contains('max')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_max", "", .x)) %>% # remove '_diff_max' from column names
  pivot_longer(
    cols = output:sundry,
    names_to = "source_of_cost_difference",
    values_to = "rotation_max",
    values_drop_na = TRUE
  )

# Join data frames
sources_of_costs <- dplyr::left_join(mean_cost_diff_df, min_cost_diff_df, by = c("region", "soiltype", "initDR", "source_of_cost_difference")) %>% 
  dplyr::left_join(., max_cost_diff_df, by = c("region", "soiltype", "initDR", "source_of_cost_difference")) %>% 
  dplyr::mutate(across('source_of_cost_difference', ~ str_replace(., 'cyl', 'lost yield'))) %>%
  dplyr::mutate(across('source_of_cost_difference', ~ str_replace(., 'herb', 'herbicide'))) %>%
  dplyr::mutate(across('source_of_cost_difference', ~ str_replace(., 'ops', 'operations'))) %>%
  dplyr::mutate(across('source_of_cost_difference', ~ str_replace(., 'fert', 'fertiliser'))) %>% 
  dplyr::mutate(across('source_of_cost_difference', ~ str_replace(., 'seed', 'seed'))) %>% 
  dplyr::mutate(across('source_of_cost_difference', ~ str_replace(., 'sundry', 'sundry'))) %>% 
  dplyr::mutate(source_of_cost_difference = factor(source_of_cost_difference,
                                                   levels = c(
                                                     "output",
                                                     "lost yield", 
                                                     "herbicide",
                                                     "fertiliser",
                                                     "operations",
                                                     "seed",
                                                     "sundry")))
  
# Tidy up
rm(mean_cost_diff_df, min_cost_diff_df, max_cost_diff_df)
```


### 6.2.3. Plot

Specify panel labels.
```{r, specify-panel-labels, echo=FALSE}
denres.labs <- c("LD-LR", "LD-HR", "HD-HR") # new
names(denres.labs) <- c("LD-LR", "LD-HR", "HD-HR") # original

soiltype.labs <- c("Heavy soil", "Medium soil", "Light soil") # new
names(soiltype.labs) <- c("heavy", "medium", "light") # original
```

Make plots & combine.  
A -ve value means MIT wins (green = MIT wins, orange = MIT loses) 
```{r}
# north ----
cost_sources_north <- ggplot(sources_of_costs %>% 
                           dplyr::filter(region == "north" & !grepl("output", source_of_cost_difference)),
               aes(source_of_cost_difference, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#CC6677"), 
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR, 
             labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs)) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of costs", 
       y = expression(atop("Mean annual difference in costs",
                           paste("between scenarios (MIT-BAU) (£ / ha)")))) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(plot.title = element_text(size = 22, face = "bold"),
        strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title = element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        #axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none"
        ) +
  ggtitle('North')


# central ----
cost_sources_central <- ggplot(sources_of_costs %>% 
                           dplyr::filter(region == "central" & !grepl("output", source_of_cost_difference)),
               aes(source_of_cost_difference, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#CC6677"), 
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of costs", 
       y = expression(atop("Mean annual difference in costs",
                           paste("between scenarios (MIT-BAU) (£ / ha)")))) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(plot.title = element_text(size = 22, face = "bold"),
        strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title = element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        #axis.text.x = element_text(size=rel(2.2)),
        axis.title.y = element_blank(),
        axis.text.y = element_text(size=rel(2.2)),
        legend.text = element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none") +
  ggtitle('Central')


# east ----
cost_sources_east <- ggplot(sources_of_costs %>% 
                           dplyr::filter(region == "east" & !grepl("output", source_of_cost_difference)),
               aes(source_of_cost_difference, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#CC6677"), 
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of costs", 
       y = expression(atop("Mean annual difference in costs",
                           paste("between scenarios (MIT-BAU) (£ / ha)")))) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(plot.title = element_text(size = 22, face = "bold"),
        strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title = element_text(size = 24),
        axis.text.x = element_text(size = rel(2.2)),
        axis.text.y = element_text(size = rel(2.2)),
        legend.text = element_text(size = 22),
        legend.title = element_blank(),
        legend.position = "bottom") +
  ggtitle('East')

# combine vertical ----
cost_sources_1col <- cost_sources_north + cost_sources_central + cost_sources_east + plot_layout(ncol = 1) +
  plot_annotation(
    tag_levels = 'a',
    #tag_levels = list(c("(a) N", "(b) C" , "(c) E")),
    tag_prefix = '(',
    tag_suffix = ')'
    ) &
  theme(plot.tag = element_text(size = 22, face="bold"))


# combine horizontal ----
# cost_sources_3col <- cost_sources_north + cost_sources_central + cost_sources_east + plot_layout(ncol = 2) +
#   plot_annotation(
#     tag_levels = 'a', 
#     #tag_levels = list(c("(a) N", "(b) C" , "(c) E")),
#     tag_prefix = '(',  
#     tag_suffix = ')'
#     ) & 
#   theme(plot.tag = element_text(size = 22, face="bold"))

# combine square ----
cost_sources_2col <- cost_sources_north + cost_sources_central + cost_sources_east + plot_layout(ncol = 2, nrow = 2) +
  plot_annotation(
    tag_levels = 'a', 
    #tag_levels = list(c("N", "C" , "E")),
    tag_prefix = '(',  
    tag_suffix = ')'
    ) & 
  theme(plot.tag = element_text(size = 22, face="bold"))
```





### 6.2.4. Save plot
```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_all_outcomes_1col.png",
#     height = 30, width = 10, units = "in", res = 300)
# print(facetpanel_1col)
# dev.off()

pdf("../figures/costs-only_sources_1col.pdf", height = 30, width = 10)
print(cost_sources_1col)
dev.off()
# 
# 
# pdf("../figures/costs-only_sources_3col.pdf", height = 10, width = 30)
# print(cost_sources_3col)
# dev.off()

pdf("../figures/costs-only_sources_2col.pdf", height = 25, width = 20)
print(cost_sources_2col)
dev.off()
```


## 6.3. Total costs and output
### 6.3.1. Calculate the differences between BAU and MIT for output and total costs
Add variable and operations costs to get total costs.  
  
Herbicide cost (£/ha) = ('bgherb' application rate (l/ha) × bgherb price (£/l)) + ('glyphosate' application rate (l/ha) × (glyphosate price (£/l))  
Fertiliser cost (£/ha) = (N fertiliser rate (kg/ha) × N Fertiliser price (£/kg)) + (P fertiliser rate (kg/ha) × P fertiliser price (£/kg)) + (K fertiliser rate (kg/ha) × K fertiliser price (£/kg))  
Seed cost (£/ha) = Seed rate (kg/ha) × Seed price (£/kg)  
Sundry cost (£/ha) is primarily the cost of other chemicals and other miscellaneous costs.  
TOTAL: *Variable cost (£/ha) = Fertiliser cost + Seed cost + Herbicide cost + Sundry cost*  
Fuel cost (£/ha) = Machine work rate (h/ha) × Fuel consumption (l/h) × Fuel price (£/l)  
Labour cost (£/ha) = Work rate (h/ha) × Hourly labour wage (£/h)  
TOTAL: *Operation cost (£/ha) = Fuel cost + Labour cost*  
  
```{r}
cost_output_diffs <- c_gp_yld %>% 

# 1. Sum variable and ops costs
  rowwise() %>% 
  dplyr::mutate(tot_costs = sum(variables_cost, operations_cost, na.rm = T)) %>%  # new category: h & vh
  ungroup() %>% 
  
# 2. Calculate difference between MIT and BAU
  dplyr::group_by(initDR, region, soiltype, year) %>% 
  dplyr::mutate(
    output_diff = output - output[scenario == "BAU"], 
    costs_diff = tot_costs - tot_costs[scenario == "BAU"]
  ) %>% 
  ungroup() %>% 

# 2. Get rotation min and max
  dplyr::filter(scenario == "MIT") %>% 
  dplyr::group_by(initDR, region, soiltype) %>% # this grouping will give range across years
  dplyr::mutate(
    # min & max difference in output
    output_diff_min = min(output_diff, na.rm = TRUE), 
    output_diff_max = max(output_diff, na.rm = TRUE), 
    # min & max difference in cost due to yield losses
    costs_diff_min = min(costs_diff, na.rm = TRUE), 
    costs_diff_max = max(costs_diff, na.rm = TRUE)
  ) %>%
  ungroup() %>%

# 3. Calculate rotation mean
  dplyr::group_by(initDR, region, soiltype) %>% 
  dplyr::mutate(
    output_diff_mean = mean(output_diff),
    costs_diff_mean = mean(costs_diff)
  ) %>% 
  ungroup() %>% 
  
# 4. Retain just useful rows/columns & tidy
  distinct(initDR, region, soiltype, .keep_all = TRUE) %>%
  dplyr::select (initDR, region, soiltype, 
                 output_diff_mean, output_diff_min, output_diff_max,
                 costs_diff_mean, costs_diff_min, costs_diff_max
                 )
```

### 6.3.2. Convert from wide to long for plotting
```{r}
mean_CO_diff <- cost_output_diffs %>%
  dplyr::select(region, soiltype, initDR, contains('mean')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_mean", "", .x)) %>% # remove '_diff_mean' from column names
  pivot_longer(
    cols = output:costs,
    names_to = "source",
    values_to = "rotation_mean",
    values_drop_na = TRUE
  )

min_CO_diff <- cost_output_diffs %>%
  dplyr::select(region, soiltype, initDR, contains('min')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_min", "", .x)) %>% # remove '_diff_min' from column names
  pivot_longer(
    cols = output:costs,
    names_to = "source",
    values_to = "rotation_min",
    values_drop_na = TRUE
  )

max_CO_diff <- cost_output_diffs %>%
  dplyr::select(region, soiltype, initDR, contains('max')) %>% 
  dplyr::rename_at(vars(everything()), ~ sub("_diff_max", "", .x)) %>% # remove '_diff_max' from column names
  pivot_longer(
    cols = output:costs,
    names_to = "source",
    values_to = "rotation_max",
    values_drop_na = TRUE
  )

# Join data frames
CO_df <- dplyr::left_join(mean_CO_diff, min_CO_diff, by = c("region", "soiltype", "initDR", "source")) %>% 
  dplyr::left_join(., max_CO_diff, by = c("region", "soiltype", "initDR", "source")) %>% 
  dplyr::mutate(across('source', ~ str_replace(., 'costs', 'total costs'))) %>% 
  dplyr::mutate(source = factor(source, levels = c("output", "total costs")))
  
# Tidy up
rm(mean_CO_diff, min_CO_diff, max_CO_diff)
```


### 6.2.3. Plot

Specify panel labels.
```{r, specify-panel-labels, echo=FALSE}
denres.labs <- c("LD-LR", "LD-HR", "HD-HR") # new
names(denres.labs) <- c("LD-LR", "LD-HR", "HD-HR") # original

soiltype.labs <- c("Heavy soil", "Medium soil", "Light soil") # new
names(soiltype.labs) <- c("heavy", "medium", "light") # original
```


Plot output and costs - without the max and min as the ranges make the points so close to zero that can't see much.  
For output, a +ve value means MIT wins (green = MIT wins, pink = MIT loses)  
For costs, a -ve value means MIT wins (green = MIT wins, pink = MIT loses)  
```{r}
rm(CO_df_2)
# Create a new column to specify colours
CO_df_2 <- CO_df %>%
  dplyr::mutate(color = case_when(
    grepl("costs", source) & rotation_mean <= 0 ~ "#009E73",  # If 'costs' and y <= 0, color green
    grepl("costs", source) & rotation_mean > 0  ~ "#CC6677",    # If 'costs' and y > 0, color red
    grepl("output", source) & rotation_mean < 0 ~ "#CC6677",    # If 'output' and y < 0, color red
    grepl("output", source) & rotation_mean >= 0 ~ "#009E73", # If 'output' and y >= 0, color green
    TRUE ~ NA_character_                # Default case, if no condition is met
  ))

# Create a new column for shape based on the conditions
CO_df_2 <- CO_df_2 %>%
  dplyr::mutate(pointshape = case_when(
    source == 'output' ~ 3,  # Cross symbol
    grepl('cost', source) ~ 16, # solid circle
    TRUE ~ 1  # Default to a Circle symbol for other cases
  ))

# North
CO_north <- ggplot(CO_df_2 %>% dplyr::filter(region == "north"),
       aes(x = source, y = rotation_mean)) +
  geom_point(aes(shape = factor(pointshape), color = color, size = 3)) +  # Shape and color aesthetics
  scale_shape_manual(values = c(3, 16)) +  # Cross (3) and Circle (16)
  scale_color_identity() +  # Use the color column directly
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs)
             )  +
  labs(x = "Output and costs", 
       y = expression(atop("Mean annual difference between",
                           paste("scenarios (MIT-BAU) (£ / ha)")))
       ) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title = element_text(size = 24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        #axis.text.x = element_text(size=rel(2.2)),
        axis.text.y = element_text(size=rel(2.2)),
        legend.text = element_text(size = 22),
        legend.title = element_blank(),
        legend.position = "none",
        tag.text = element_text(size = 22)
        #strip.placement = 'outside',
        )

# Central
CO_central <- ggplot(CO_df_2 %>% dplyr::filter(region == "central"),
       aes(x = source, y = rotation_mean)) +
  geom_point(aes(shape = factor(pointshape), color = color, size = 3)) +  # Shape and color aesthetics
  scale_shape_manual(values = c(3, 16)) +  # Cross (3) and Circle (1)
  scale_color_identity() +  # Use the color column directly
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs)
             )  +
  labs(x = "Output and costs", 
       y = expression(atop("Mean annual difference between",
                           paste("scenarios (MIT-BAU) (£ / ha)")))
       ) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title = element_text(size = 24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        #axis.text.x = element_text(size=rel(2.2)),
        axis.text.y = element_text(size=rel(2.2)),
        legend.text = element_text(size = 22),
        legend.title = element_blank(),
        legend.position = "none",
        tag.text = element_text(size = 22)#,
        #strip.placement = 'outside',
        )

# East
CO_east <- ggplot(CO_df_2 %>% dplyr::filter(region == "east"),
       aes(x = source, y = rotation_mean)) +
  geom_point(aes(shape = factor(pointshape), color = color, size = 3)) +  # Shape and color aesthetics
  scale_shape_manual(values = c(3, 16)) +  # Cross (3) and Circle (1)
  scale_color_identity() +  # Use the color column directly
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs)
             )  +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Income (farm output) and expenditure (costs)", 
       y = expression(atop("Mean annual difference between",
                           paste("scenarios (MIT-BAU) (£ / ha)")))
       ) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        #axis.title.x = element_blank(),
        #axis.text.x = element_blank(),
        axis.text.x = element_text(size=rel(2.2)),
        axis.text.y = element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none",
        tag.text = element_text(size=22)#,
        #strip.placement = 'outside',
        )


# Combine
CO_1col <- CO_north + CO_central + CO_east + plot_layout(ncol = 1) +
  plot_annotation(
    tag_levels = list(c("(a) North", "(b) Central" , "(c) East")),
    tag_prefix = '',  
    tag_suffix = ''
    ) & 
  theme(plot.tag = element_text(size = 22, face="bold"))
```





### 6.3.4. Save plot
```{r, echo=FALSE, eval = FALSE}
# png("../figures/??.png",
#     height = 30, width = 10, units = "in", res = 300)
# print(facetpanel_1col)
# dev.off()

pdf("../figures/costs_and_output_1col.pdf", height = 30, width = 10)
print(CO_1col)
dev.off()
```








# Cite packages
```{r}
cite_packages(citation.style = "nature", out.dir = ".")
```
See here https://pakillo.github.io/grateful/index.html for instructions on using `grateful`.



# Superseded
## 3.3. NOT USED - Calculate % change from individual years
Here, I calculate the difference between BAU and MIT for each year, then calculate the % change for each year, then take the average % change.
```{r}
# pc_change <- c_gp_yld %>%
#   
#   # 1. Calculate percentage change for each year of each scenario
#   dplyr::group_by(region, soiltype, initDR, year) %>% 
#   dplyr::mutate(
#     
#     # gross profit % change
#     gp_pc_change = (oc/gp[scenario == "BAU"])*100, # opportunity cost is already the difference in gross profit 
#     
#     # calories % change
#     cal_pc_change = (calories_diff/calories[scenario == "BAU"])*100,    
#     
#     # land use % change
#     lu_pc_change = (land_use_diff/land_use[scenario == "BAU"])*100,  
#     
#     # carbon emissions % change
#     c_pc_change = (c_emiss_diff/c_total[scenario == "BAU"])*100
#     
#   ) %>% 
#   ungroup() %>% 
#   
#   # 2. calculate min and max values (this grouping will give range across years)
#   dplyr::group_by(scenario, region, soiltype, initDR, year) %>% 
#   dplyr::mutate(
#     
#     # gross profit difference min & max
#     gp_diff_min = min(oc, na.rm = TRUE),
#     gp_diff_max = max(oc, na.rm = TRUE),
#     
#     # calories difference min & max
#     calories_diff_min = min(calories_diff, na.rm = TRUE),
#     calories_diff_max = max(calories_diff, na.rm = TRUE),
#     
#     # land use difference min & max
#     land_use_diff_min = min(land_use_diff, na.rm = TRUE),
#     land_use_diff_max = max(land_use_diff, na.rm = TRUE),
#     
#     # GHG difference min & max
#     c_emiss_diff_min = min(c_emiss_diff, na.rm = TRUE),
#     c_emiss_diff_max = max(c_emiss_diff, na.rm = TRUE)
#     
#   ) %>% 
#   ungroup() %>% 
#   
#   # 3. Calculate mean for each level of initDR
#   dplyr::filter(scenario == "MIT") %>% 
#   dplyr::group_by(initDR) %>% 
#   dplyr::mutate(
#     # gross profit mean, min, max % change
#     gp_pcchange_mean = mean(gp_pc_change),
#     gp_pcchange_min = min(gp_pc_change),
#     gp_pcchange_max = max(gp_pc_change),
#     
#     # calories mean min, max % change
#     cal_pcchange_mean = mean(cal_pc_change),
#     cal_pcchange_min = min(cal_pc_change),
#     cal_pcchange_max = max(cal_pc_change),
#     
#     # land use mean min, max % change
#     lu_pcchange_mean = mean(lu_pc_change),
#     lu_pcchange_min = min(lu_pc_change),
#     lu_pcchange_max = max(lu_pc_change),
#     
#     # carbon emissions mean min, max % change
#     c_pcchange_mean = mean(c_pc_change),
#     c_pcchange_min = min(c_pc_change),
#     c_pcchange_max = max(c_pc_change)
#     
#   ) %>% 
#   
#   #Select just useful columns and filter out duplicates
#   dplyr::select(initDR, contains("pcchange")) %>% 
#   dplyr::distinct(initDR, .keep_all = TRUE) %>% 
#   arrange(initDR)
# 
# pc_change
```

## 3.4. Save output as .csv
```{r}
#write.csv(pc_change, "../output/summarytable_from_indiv_years.csv", row.names = FALSE)
```

## 6.1. Yield plot 

```{r}
# Specify panel labels
region.labs <- c("Northern England", "Central England", "Eastern England") # new
names(region.labs) <- c("north", "central", "east") # original

soiltype.labs <- c("Heavy soil", "Medium soil", "Light soil") # new
names(soiltype.labs) <- c("heavy", "medium", "light") # original

# Specify plot
#library(ggbreak)
figSx_yield <- ggplot(trdoff_rot, aes(initDR, yield_mean, colour = scenario, shape = scenario)) + 
  facet_grid(region ~ soiltype,
             labeller = labeller(region = region.labs, 
                                 soiltype = soiltype.labs)) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  ylim(0, 10) +
  labs(x = "Initial density and resistance of Black-grass population", 
       y = "Mean annual crop yield (t/ha)") +
  # specify colours for 'scenario':
  scale_color_manual(name = "Scenario",
                     values = c("#00306F", # BAU, madison blue
                                "#00C2F9") # MIT, capri blue
                     ) + 
  scale_shape_manual(name = "Scenario",
                     values=c(15, # BAU, 15 = square
                              16) # MIT, 16 = circle
                     ) + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.title=element_text(size=24),
        legend.text=element_text(size=24)
        ) +
  geom_pointrange(aes(ymin = yield_min, ymax = yield_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9)

figSx_yield #+ scale_y_break(c(20, 60))
```

Save.
```{r}
png("../figures/Figure_Sx_yield-t-ha.png",
    height = 10, width = 10, units = "in", res = 300)
print(figSx_yield)
dev.off()
```


```{r}
library(ggbreak)
E_L <- ggplot(trdoff_rot %>% dplyr::filter(region == "east" & soiltype == "light"), 
       aes(initDR, yield_mean, colour = scenario, shape = scenario)) + 
  #facet_grid(region ~ soiltype,
  #           labeller = labeller(region = region.labs, 
  #                               soiltype = soiltype.labs)) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Initial density and resistance of Black-grass population", 
       y = "Mean annual crop yield (t/ha)") +
  # specify colours for 'scenario':
  scale_color_manual(name = "Scenario",
                     values = c("#00306F", # BAU, madison blue
                                "#00C2F9") # MIT, capri blue
                     ) + 
  scale_shape_manual(name = "Scenario",
                     values=c(15, # BAU, 15 = square
                              16) # MIT, 16 = circle
                     ) + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.title=element_text(size=24),
        legend.text=element_text(size=24)
        ) +
  geom_pointrange(aes(ymin = yield_min, ymax = yield_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) + 
  #scale_y_break(c(20, 60))
  #scale_y_break(c(20, 60), space = 0.1, ticklabels = c(seq(0,80,20), 70), expand = c(0, 0))
  #scale_y_break(c(16, 67), scales = 0.5) # THIS WORKS OK ISH
  # this works better:
  scale_y_break(c(8, 15), ticklabels = c(seq(15,16,1), 16), scales = 0.3) + 
  scale_y_break(c(16, 67), ticklabels = c(seq(67,69,1), 70), scale=0.5) +
  theme(axis.text.x.top = element_blank(),
        axis.ticks.x.top = element_blank(),
        axis.line.x.top = element_blank())
  #scale_y_cut(breaks=c(8, 20), which=c(1, 3), scales=c(0.5, 3))

E_L
```




## 7.alt Mean difference in C emissions per initDR
This gives min and max values for any single year of a management strategy. *This is NOT what I use in the paper.*  
```{r}
c_gp_yld %>%
  dplyr::filter(scenario == 'MIT') %>% 
  dplyr::group_by(initDR) %>% 
  dplyr::summarize(
    #gp_mean = mean(gp),
    #cal_mean = mean(calories),
    c_emiss_meandiff = mean(c_emiss_diff),
    c_emiss_min = min(c_emiss_diff),
    c_emiss_max = max(c_emiss_diff)
  ) %>% 
  ungroup() 
``` 
### 7.2.2. Difference in gross profit
+ve value means MIT wins (green = MIT wins, orange = MIT loses)  
```{r, echo=FALSE}
facet_oc <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
               aes(initDR, oc_mean, colour = oc_mean < 0)) + 
  scale_colour_manual(values = c("#009E73", "#D55E00"),
                      #name = "MIT wins?",
                      labels = c("MIT wins", "MIT loses")) +
                      #labels = c("Yes", "No")) +
  facet_grid(soiltype ~ region,
             labeller = labeller(soiltype = soiltype.labs, 
                                 region = region.labs)#,
             #scales="free_y"
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Initial density and resistance of Black-grass population", 
       y = "Mean annual difference in gross profit (£/ha)" 
       ) +
  geom_pointrange(aes(ymin = oc_min, ymax = oc_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank()
        )

facet_oc
```
  
  

```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_oc.png",
#     height = 15, width = 15, units = "in", res = 300)
# print(facet_oc)
# dev.off()

pdf("../figures/plot_oc.pdf", height = 15, width = 15)
print(facet_oc)
dev.off()
```



### 7.2.3. Difference in yield (calories)
+ve value means MIT wins
```{r, echo=FALSE}
facet_cal <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
               aes(initDR, calories_diff_mean, colour = calories_diff_mean < 0)) + 
  scale_colour_manual(values = c("#009E73", "#D55E00"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ region,
             labeller = labeller(soiltype = soiltype.labs, 
                                 region = region.labs)#,
             #scales="free_y"
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Initial density and resistance of Black-grass population", 
       y = "Mean annual difference in calories produced (kcal/ha)" 
       ) +
  geom_pointrange(aes(ymin = calories_diff_min, ymax = calories_diff_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank()
        )

facet_cal
```
  
  

```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_cal.png",
#     height = 15, width = 15, units = "in", res = 300)
# print(facet_cal)
# dev.off()

pdf("../figures/plot_cal.pdf", height = 15, width = 15)
print(facet_cal)
dev.off()
```



### 7.2.4. Difference in GHG emissions
-ve value means MIT wins
```{r, echo=FALSE}
facet_c <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
               aes(initDR, c_emiss_diff_mean, colour = c_emiss_diff_mean > 0)) + 
    scale_colour_manual(values = c("#009E73", "#D55E00"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ region,
             labeller = labeller(soiltype = soiltype.labs, 
                                 region = region.labs)#,
             #scales="free_y"
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Initial density and resistance of Black-grass population", 
       y = expression("Mean annual difference in GHG emissions (kgCO"["2"]~"eq/ha)") 
       ) +
  geom_pointrange(aes(ymin = c_emiss_diff_min, ymax = c_emiss_diff_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank()
        )
facet_c
```
  
  

```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_GHG.png",
#     height = 15, width = 15, units = "in", res = 300)
# print(facet_c)
# dev.off()

pdf("../figures/plot_GHG.pdf", height = 15, width = 15)
print(facet_c)
dev.off()
```




### 7.2.5. alt - Square
```{r}
# # opp costs ----
# a_oc <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
#                aes(initDR, oc_mean, colour = oc_mean < 0)) + 
#   scale_colour_manual(values = c("#009E73", "#D55E00"),
#                       labels = c("MIT wins", "MIT loses")) +
#   facet_grid(soiltype ~ region,
#              labeller = labeller(soiltype = soiltype.labs, region = region.labs)) +
#   scale_x_discrete(guide = guide_axis(angle = 45)) +
#   labs(x = "Initial density and resistance of Black-grass population", 
#        y = "Mean annual difference in gross profit (£/ha)") +
#   geom_pointrange(aes(ymin = oc_min, ymax = oc_max), 
#                   position=position_dodge(width=0.3),
#                   size = 0.9) +
#   geom_hline(yintercept=0, linetype="dashed", color = "black") +
#   theme_bw() + 
#   theme(strip.placement = 'outside',
#         strip.text.x = element_text(size = 22),
#         strip.text.y = element_text(size = 22),
#         axis.title=element_text(size=24),
#         axis.text.x=element_text(size=rel(2.2)),
#         axis.text.y=element_text(size=rel(2.2)),
#         legend.text=element_text(size=22),
#         legend.title = element_blank(),
#         legend.position = "none"
#         ) 
# 
# # yield ----
# b_cal <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
#                aes(initDR, calories_diff_mean, colour = calories_diff_mean < 0)) + 
#   scale_colour_manual(values = c("#009E73", "#D55E00"),
#                       labels = c("MIT wins", "MIT loses")) +
#   facet_grid(soiltype ~ region,
#              labeller = labeller(soiltype = soiltype.labs, region = region.labs)) +
#   scale_x_discrete(guide = guide_axis(angle = 45)) +
#   labs(x = "Initial density and resistance of Black-grass population", 
#        y = "Mean annual difference in calories produced (kcal/ha)") +
#   geom_pointrange(aes(ymin = calories_diff_min, ymax = calories_diff_max), 
#                   position=position_dodge(width=0.3),
#                   size = 0.9) +
#   geom_hline(yintercept=0, linetype="dashed", color = "black") +
#   theme_bw() + 
#   theme(strip.placement = 'outside',
#         strip.text.x = element_text(size = 22),
#         strip.text.y = element_text(size = 22),
#         axis.title=element_text(size=24),
#         axis.text.x=element_text(size=rel(2.2)),
#         axis.text.y=element_text(size=rel(2.2)),
#         legend.text=element_text(size=22),
#         legend.title = element_blank(),
#         legend.position = "none"
#         )
# 
# # GHG emissions ----
# c_ghg <- ggplot(trdoff_rot %>% dplyr::filter(scenario == "MIT"), 
#                aes(initDR, c_emiss_diff_mean, colour = c_emiss_diff_mean > 0)) + 
#     scale_colour_manual(values = c("#009E73", "#D55E00"),
#                       labels = c("MIT wins", "MIT loses")) +
#   facet_grid(soiltype ~ region,
#              labeller = labeller(soiltype = soiltype.labs, 
#                                  region = region.labs)) +
#   scale_x_discrete(guide = guide_axis(angle = 45)) +
#   labs(x = "Initial density and resistance of Black-grass population", 
#        y = expression("Mean annual difference in GHG emissions (kgCO"["2"]~"eq/ha)")) +
#   geom_pointrange(aes(ymin = c_emiss_diff_min, ymax = c_emiss_diff_max), 
#                   position=position_dodge(width=0.3),
#                   size = 0.9) +
#   geom_hline(yintercept=0, linetype="dashed", color = "black") +
#   theme_bw() + 
#   theme(strip.placement = 'outside',
#         strip.text.x = element_text(size = 22),
#         strip.text.y = element_text(size = 22),
#         axis.title=element_text(size=24),
#         axis.text.x=element_text(size=rel(2.2)),
#         axis.text.y=element_text(size=rel(2.2)),
#         legend.text=element_text(size=22),
#         legend.title = element_blank(),
#         legend.position = "none"
#         ) 
# 
# # get legend
# g_legend<-function(a.gplot){
#     tmp <- ggplot_gtable(ggplot_build(a.gplot))
#     leg <- which(sapply(tmp$grobs, function(x) x$name) == "guide-box")
#     legend <- tmp$grobs[[leg]]
#     legend
# }
# 
# legend <- g_legend(facet_oc)
# 
# # combine
# facetpanel <- a_oc + b_cal + c_ghg + legend + plot_layout(ncol = 2)
```



```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_all_outcomes_square.png",
#     height = 20, width = 20, units = "in", res = 300)
# print(facetpanel)
# dev.off()

# pdf("../figures/plot_all_outcomes_square.pdf", height = 20, width = 20)
# print(facetpanel)
# dev.off()
```


### 7.3.2. Plot north region 
-ve value means MIT wins (green = MIT wins, orange = MIT loses)  
```{r, echo=FALSE}
eissions_sources_north <- ggplot(sources %>% dplyr::filter(region == "north"), 
               aes(source_of_emissions, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#D55E00"),
                      #name = "MIT wins?",
                      labels = c("MIT wins", "MIT loses")) +
                      #labels = c("Yes", "No")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)#,
             #scales="free_y"
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of GHG emissions", 
       y = expression("Mean annual difference in GHG emissions (kgCO"["2"]~"eq / ha)") 
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank()
        )

eissions_sources_north
```
 


### 7.3.3. Plot central region 
-ve value means MIT wins (green = MIT wins, orange = MIT loses)  
```{r, echo=FALSE}
eissions_sources_central <- ggplot(sources %>% dplyr::filter(region == "central"), 
               aes(source_of_emissions, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#D55E00"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of GHG emissions", 
       y = expression("Mean annual difference in GHG emissions (kgCO"["2"]~"eq / ha)") 
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank()
        )

eissions_sources_central
```
 


### 7.3.4. Plot east region 
-ve value means MIT wins (green = MIT wins, orange = MIT loses)  
```{r, echo=FALSE}
eissions_sources_east <- ggplot(sources %>% dplyr::filter(region == "east"), 
               aes(source_of_emissions, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#009E73", "#D55E00"),
                      labels = c("MIT wins", "MIT loses")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of GHG emissions", 
       y = expression("Mean annual difference in GHG emissions (kgCO"["2"]~"eq / ha)") 
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank()
        )

eissions_sources_east
```
  

## 7.5. Diversification plots

```{r}
unique(strat$design_element)
```

Plot some of the yearly design elements.
```{r}
strat %>%
  dplyr::mutate(initdenres = fct_recode(initdenres, "BAU" = "NA")) %>% 
  dplyr::mutate(initdenres = factor(initdenres, levels=c("BAU", "LD-LR", "LD-HR", "HD-HR"))) %>%
  dplyr::mutate(ID = paste(scenario, initdenres)) %>% 
  dplyr:: group_by(ID) %>% 
  dplyr::filter(design_element %in% c("tillage_depth_yr", "herb_applic_n_yr", "selherb_vol_yr", "gly_vol_yr" )) %>% #"aut_gly_n_yr", 
  dplyr::mutate(design_element = factor(design_element, levels=c("tillage_depth_yr", "herb_applic_n_yr", "selherb_vol_yr",  "gly_vol_yr"))) %>% #"aut_gly_n_yr",
  ggplot(aes(fill=design_element, y=scaled_value, x=initdenres)) +
    #geom_point(color="lightgrey", size=1, width=0.1, position=position_jitterdodge() , alpha=0.4) +
    geom_boxplot(position=position_dodge2(preserve = "total"), alpha=0.5, outlier.colour="grey") + #, outlier.colour="transparent"
    scale_fill_viridis(discrete=T, name="",
                       labels = c("Tillage depth", 
                                  "Herbicide application frequency", 
                                  "Selective herbicide volume", 
                                  #"Autumn glyphosate application frequency",  
                                  "Glyphosate volume")
                       ) +
    #geom_text(data=strat, aes(label=paste0("n: ",n), y=median-2), position=position_dodge(1), hjust=0.5) +
    #theme_ipsum()  +
    theme_classic() +
    xlab("") +
    ylab("Scaled value") #+
    #ylim(0,1)
```


Annual volumes of herbicide applied.
```{r}
strat %>%
  dplyr::mutate(initdenres = fct_recode(initdenres, "BAU" = "NA")) %>% 
  dplyr::mutate(initdenres = factor(initdenres, levels=c("BAU", "LD-LR", "LD-HR", "HD-HR"))) %>%
  dplyr::mutate(ID = paste(scenario, initdenres)) %>% 
  dplyr:: group_by(ID) %>% 
  dplyr::filter(design_element %in% c("selherb_vol_yr", "gly_vol_yr" )) %>% 
  dplyr::mutate(design_element = factor(design_element, levels=c("selherb_vol_yr",  "gly_vol_yr"))) %>% 
  ggplot(aes(fill=design_element, y=value, x=initdenres)) +
    #geom_point(color="lightgrey", size=1, width=0.1, position=position_jitterdodge() , alpha=0.4) +
    geom_boxplot(position=position_dodge2(preserve = "total"), alpha=0.5, outlier.colour="grey") + 
    scale_fill_manual(values=c("#332288", "#117733"),
                      labels = c("Selective herbicide", 
                                  "Glyphosate")) +
    # scale_fill_manual(values=c("#999999", "#E69F00"),
    #                   labels = c("Selective herbicide", 
    #                               "Glyphosate")) + #, "#56B4E9"
    # scale_fill_viridis(discrete=T, 
    #                    #option = "D", #(this is the)
    #                    name="",
    #                    labels = c("Selective herbicide", 
    #                               "Glyphosate")
    #                    ) +
    #geom_text(data=strat, aes(label=paste0("n: ",n), y=median-2), position=position_dodge(1), hjust=0.5) +
    #theme_ipsum()  +
    theme_classic() +
    theme(legend.title=element_blank()) +
    xlab("") +
    ylab(expression("Volume "*("l "*ha^{-1}*yr^{-1})))
    #ylab(bquote('Volume'~(l^-1)))
    #ylab("Volume (l/ ha/ yr)") #+
    #ylim(0,1)
```

This plot is good as it shows, compared to BAU, sometimes reduced selective herbicide volume but increased glyphosate volume.


Rotation herbicide volumes.
```{r}
strat %>%
  dplyr::mutate(initdenres = fct_recode(initdenres, "BAU" = "NA")) %>% 
  dplyr::mutate(initdenres = factor(initdenres, levels=c("BAU", "LD-LR", "LD-HR", "HD-HR"))) %>%
  dplyr::mutate(ID = paste(scenario, initdenres)) %>% 
  dplyr:: group_by(ID) %>% 
  dplyr::filter(design_element %in% c("selherb_rot_vol", "gly_rot_vol" )) %>% 
  dplyr::mutate(design_element = factor(design_element, levels=c("selherb_rot_vol",  "gly_rot_vol"))) %>% 
  ggplot(aes(fill=design_element, y=value, x=initdenres)) +
    #geom_point(color="lightgrey", size=1, width=0.1, position=position_jitterdodge() , alpha=0.4) +
    geom_boxplot(position=position_dodge2(preserve = "total"), alpha=0.5, outlier.colour="grey") + 
    scale_fill_manual(values=c("#332288", "#117733"),
                      labels = c("Selective herbicide", 
                                  "Glyphosate")) +
    theme_classic() +
    theme(legend.title=element_blank()) +
    xlab("") +
    ylab(expression("Volume "*("l "*ha^{-1})))
```



Plot rotation active ingredients and stale seed beds.
```{r}
strat %>%
  dplyr::mutate(initdenres = fct_recode(initdenres, "BAU" = "NA")) %>% 
  dplyr::mutate(initdenres = factor(initdenres, levels=c("BAU", "LD-LR", "LD-HR", "HD-HR"))) %>%
  dplyr::mutate(ID = paste(scenario, initdenres)) %>% 
  dplyr:: group_by(ID) %>% 
  dplyr::filter(design_element %in% c("herb_applic_n_rot", "diff_act_tot_n_rot", "ssb_tot_n_rot")) %>%
  dplyr::mutate(design_element = factor(design_element, levels=c("herb_applic_n_rot", "diff_act_tot_n_rot", "ssb_tot_n_rot"))) %>% 
  ggplot(aes(fill=design_element, y=value, x=initdenres)) +
    geom_boxplot(position=position_dodge2(preserve = "total"), alpha=0.5, outlier.colour="grey") + #, outlier.colour="transparent"
    scale_fill_manual(values=c("#CC6677", "#DDCC77", "#AA4499"), #"#AA4499", "#443F41"
                      labels = c(
                        "Herbicide applications",
                        "Active ingredients",
                        "Stale seedbed")
                      ) +
    # scale_fill_viridis(discrete=T, name="",
    #                    labels = c("Number of different actives", 
    #                               "Number of different crops", 
    #                               "Number of spring crops", 
    #                               "Number of stale seed beds"#,  
    #                               #"Glyphosate total volume"
    #                               )
    #                    ) +
    geom_point(color="lightgrey", size=1, width=0.1, position=position_jitterdodge() , alpha=0.4) +
    #geom_text(data=strat, aes(label=paste0("n: ",n), y=median-2), position=position_dodge(1), hjust=0.5) +
    #theme_ipsum()  +
    theme_classic() +
    theme(legend.title=element_blank()) +
    xlab("") +
    ylab("Count (total number in rotation)") #+
    #ylim(0,1)
```

This doesn't really work. Which is the most interesting?

Plot rotation crop diversity.  
Number of spring crops  
Number of different crops  
```{r}
strat %>%
  dplyr::mutate(initdenres = fct_recode(initdenres, "BAU" = "NA")) %>% 
  dplyr::mutate(initdenres = factor(initdenres, levels=c("BAU", "LD-LR", "LD-HR", "HD-HR"))) %>%
  dplyr::mutate(ID = paste(scenario, initdenres)) %>% 
  dplyr:: group_by(ID) %>% 
  dplyr::filter(design_element %in% c("diff_crops_tot_n_rot", "spring_crops_tot_n_rot")) %>% 
  dplyr::mutate(design_element = factor(design_element, levels=c("diff_crops_tot_n_rot", "spring_crops_tot_n_rot"))) %>% 
  ggplot(aes(fill=design_element, y=value, x=initdenres)) +
    
    geom_boxplot(position=position_dodge2(preserve = "total"), alpha=0.5, outlier.colour="grey") + #, outlier.colour="transparent"
    scale_fill_manual(values=c("#88CCEE", "#44AA99"),
                       labels = c("Different crops",
                                  "Spring crops")
                       ) +
    # scale_fill_viridis(discrete=T, name="",
    #                    labels = c("n spring crops", 
    #                               "n different crops")
    #                    ) +
    #geom_text(data=strat, aes(label=paste0("n: ",n), y=median-2), position=position_dodge(1), hjust=0.5) +
    #theme_ipsum()  +
    geom_point(color="lightgrey", size=1, width=0.1, position=position_jitterdodge() , alpha=0.4) +
  
    theme_classic() +
    theme(legend.title=element_blank()) +
    xlab("") +
    ylab("Count (total number in rotation)") #+
    #ylim(0,1)
```


Plot frequencies.  
Number of herbicide applications per year.
Number of actives per year.  
Number autumn glyphosate applications per year.
```{r}
strat %>%
  dplyr::mutate(initdenres = fct_recode(initdenres, "BAU" = "NA")) %>% 
  dplyr::mutate(initdenres = factor(initdenres, levels=c("BAU", "LD-LR", "LD-HR", "HD-HR"))) %>%
  dplyr::mutate(ID = paste(scenario, initdenres)) %>% 
  dplyr:: group_by(ID) %>% 
  dplyr::filter(design_element %in% c("herb_applic_n_yr", "actives_n_yr")) %>% 
  dplyr::mutate(design_element = factor(design_element, levels=c("herb_applic_n_yr", "actives_n_yr"))) %>% 
  ggplot(aes(fill=design_element, y=value, x=initdenres)) +
    #geom_point(color="lightgrey", size=1, width=0.1, position=position_jitterdodge() , alpha=0.4) +
    geom_boxplot(position=position_dodge2(preserve = "total"), alpha=0.5, outlier.colour="grey") + #, outlier.colour="transparent"
    scale_fill_manual(values=c("#DDCC77", "#CC6677"),
                       labels = c("Herbicide applications",
                                  "Active ingredients")
                       ) +
    # scale_fill_viridis(discrete=T, name="",
    #                    labels = c("Herbicide applications", 
    #                               "Active ingredients", 
    #                               "Autumn glyphosate",
    #                               "Stale seedbeds")
    #                    ) +
    #geom_text(data=strat, aes(label=paste0("n: ",n), y=median-2), position=position_dodge(1), hjust=0.5) +
    #theme_ipsum()  +
    theme_classic() +
    theme(legend.title=element_blank()) +
    xlab("") +
    ylab("Application frequency (number / year)") +
    scale_y_continuous(breaks=seq(0,8,2))
    #ylim(0,8)
```


This one maybe not very useful. I think the number of herbicide applications rose because there were more stale seedbeds. Might be better just to show that.  
Active ingredients per year doesn't change much overall (mean goes down with increasing BG severity), but earlier plot shows more actives used across a rotation. So farmers use slightly fewer actives per crop but more different crops, which allows more different actives.  


```{r}
unique(strat$design_element)
```


## Costs and output plots
Now plot output and total costs for comparison:
A -ve value means MIT wins (green = MIT wins, orange = MIT loses) 
```{r}
# north ----
output_north <- ggplot(sources_of_costs %>% 
                           dplyr::filter(region == "north" & grepl("output", source_of_cost_difference)),
               aes(source_of_cost_difference, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#CC6677", "#009E73"), 
                      labels = c("MIT loses", "MIT wins")) +
  facet_grid(soiltype ~ initDR, 
             labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs)
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of costs", 
       y = expression(atop("Mean annual difference in costs",
                           paste("between scenarios (MIT-BAU) (£ / ha)")))
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none",
        tag.text = element_text(size=22)
        )


# central ----
output_central <- ggplot(sources_of_costs %>% 
                           dplyr::filter(region == "central" & grepl("output", source_of_cost_difference)),
               aes(source_of_cost_difference, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#CC6677", "#009E73"), 
                      labels = c("MIT loses", "MIT wins")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of costs", 
       y = expression(atop("Mean annual difference in costs",
                           paste("between scenarios (MIT-BAU) (£ / ha)")))
       #y = expression("Difference in costs between scenarios (MIT-BAU) (£ / ha)")
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "none"
        ) 


# east ----
output_east <- ggplot(sources_of_costs %>% 
                           dplyr::filter(region == "east" & grepl("output", source_of_cost_difference)),
               aes(source_of_cost_difference, rotation_mean, colour = rotation_mean > 0)) + 
  scale_colour_manual(values = c("#CC6677", "#009E73"), 
                      labels = c("MIT loses", "MIT wins")) +
  facet_grid(soiltype ~ initDR,
             labeller = labeller(soiltype = soiltype.labs, 
                                 initDR = denres.labs)
             ) +
  scale_x_discrete(guide = guide_axis(angle = 45)) +
  labs(x = "Source of costs", 
       y = expression(atop("Mean annual difference in costs",
                           paste("between scenarios (MIT-BAU) (£ / ha)")))
       #y = expression("Difference in costs between scenarios (MIT-BAU) (£ / ha)")
       ) +
  geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
                  position=position_dodge(width=0.3),
                  size = 0.9) +
  geom_hline(yintercept=0, linetype="dashed", color = "black") +
  theme_bw() + 
  theme(strip.placement = 'outside',
        strip.text.x = element_text(size = 22),
        strip.text.y = element_text(size = 22),
        axis.title=element_text(size=24),
        axis.text.x=element_text(size=rel(2.2)),
        axis.text.y=element_text(size=rel(2.2)),
        legend.text=element_text(size=22),
        legend.title = element_blank(),
        legend.position = "bottom"
        ) 



# combine
output_1col <- output_north + output_central + output_east + plot_layout(ncol = 1) +
  plot_annotation(
    #tag_levels = 'a', 
    tag_levels = list(c("(a) North", "(b) Central" , "(c) East")),
    tag_prefix = '',  
    tag_suffix = ''
    ) & 
  theme(plot.tag = element_text(size = 22, face="bold"))
```

Save plot.
```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_all_outcomes_1col.png",
#     height = 30, width = 10, units = "in", res = 300)
# print(facetpanel_1col)
# dev.off()

pdf("../figures/output_1col.pdf", height = 30, width = 10)
print(output_1col)
dev.off()
```






Plot again but include output this time.
For output, a +ve value means MIT wins.
For costs, a -ve value means MIT wins (green = MIT wins, orange = MIT loses) 
```{r}
# # Specify colours
# sources_of_costs <- sources_of_costs %>%
#   dplyr::mutate(color = case_when(
#     grepl("costs", source_of_cost_difference) & rotation_mean <= 0 ~ "#009E73",  # If 'costs' and y <= 0, color green
#     grepl("costs", source_of_cost_difference) & rotation_mean > 0  ~ "#CC6677",    # If 'costs' and y > 0, color red
#     grepl("output", source_of_cost_difference) & rotation_mean < 0 ~ "#CC6677",    # If 'output' and y < 0, color red
#     grepl("output", source_of_cost_difference) & rotation_mean >= 0 ~ "#009E73", # If 'output' and y >= 0, color green
#     TRUE ~ NA_character_                # Default case, if no condition is met
#   ))
# 
# 
# 
# # north ----
# costoutput_sources_n_1col <- ggplot(sources_of_costs %>% 
#                            dplyr::filter(region == "north"),
#                aes(source_of_cost_difference, rotation_mean, colour = color)) + 
#   scale_colour_identity() +
#   facet_grid(soiltype ~ initDR, 
#              labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs)#,
#              #scales="free_y"
#              ) +
#   scale_x_discrete(guide = guide_axis(angle = 45)) +
#   labs(x = "Source of costs", 
#        y = expression(atop("Mean annual difference in output",
#                            paste("and costs between scenarios (MIT-BAU) (£ / ha)"))) 
#        #y = expression("Difference in output and costs between scenarios (MIT-BAU) (£ / ha)")
#        ) +
#   geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
#                   position=position_dodge(width=0.3),
#                   size = 0.9) +
#   geom_hline(yintercept=0, linetype="dashed", color = "black") +
#   theme_bw() + 
#   theme(strip.placement = 'outside',
#         strip.text.x = element_text(size = 22),
#         strip.text.y = element_text(size = 22),
#         axis.title=element_text(size=24),
#         axis.title.x = element_blank(),
#         axis.text.x = element_blank(),
#         axis.text.y=element_text(size=rel(2.2)),
#         legend.text=element_text(size=22),
#         legend.title = element_blank(),
#         legend.position = "none",
#         tag.text = element_text(size=22)
#         )
# 
# 
# # central ----
# costoutput_sources_c_1col <- ggplot(sources_of_costs %>% 
#                            dplyr::filter(region == "central"),
#                aes(source_of_cost_difference, rotation_mean, colour = color)) + 
#   scale_colour_identity() +
#   facet_grid(soiltype ~ initDR,
#              labeller = labeller(soiltype = soiltype.labs, 
#                                  initDR = denres.labs)
#              ) +
#   scale_x_discrete(guide = guide_axis(angle = 45)) +
#   labs(x = "Source of costs", 
#        y = expression(atop("Mean annual difference in output",
#                            paste("and costs between scenarios (MIT-BAU) (£ / ha)"))) 
#        #y = expression("Difference in output and costs between scenarios (MIT-BAU) (£ / ha)")
#        ) +
#   geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
#                   position=position_dodge(width=0.3),
#                   size = 0.9) +
#   geom_hline(yintercept=0, linetype="dashed", color = "black") +
#   theme_bw() + 
#   theme(strip.placement = 'outside',
#         strip.text.x = element_text(size = 22),
#         strip.text.y = element_text(size = 22),
#         axis.title=element_text(size=24),
#         axis.title.x = element_blank(),
#         axis.text.x = element_blank(),
#         axis.text.y=element_text(size=rel(2.2)),
#         legend.text=element_text(size=22),
#         legend.title = element_blank(),
#         legend.position = "none"
#         ) 
# 
# 
# # east ----
# costoutput_sources_e_1col <- ggplot(sources_of_costs %>% 
#                            dplyr::filter(region == "east"),
#                aes(source_of_cost_difference, rotation_mean, colour = color)) + 
#   scale_colour_identity() +
#   facet_grid(soiltype ~ initDR,
#              labeller = labeller(soiltype = soiltype.labs, 
#                                  initDR = denres.labs)
#              ) +
#   scale_x_discrete(guide = guide_axis(angle = 45)) +
#   labs(x = "Source of costs", 
#        y = expression(atop("Mean annual difference in output",
#                            paste("and costs between scenarios (MIT-BAU) (£ / ha)"))) 
#        #y = expression("Difference in output and costs between scenarios (MIT-BAU) (£ / ha)")
#        ) +
#   geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max), 
#                   position=position_dodge(width=0.3),
#                   size = 0.9) +
#   geom_hline(yintercept=0, linetype="dashed", color = "black") +
#   theme_bw() + 
#   theme(strip.placement = 'outside',
#         strip.text.x = element_text(size = 22),
#         strip.text.y = element_text(size = 22),
#         axis.title=element_text(size=24),
#         axis.text.x=element_text(size=rel(2.2)),
#         axis.text.y=element_text(size=rel(2.2)),
#         legend.text=element_text(size=22),
#         legend.title = element_blank(),
#         legend.position = "bottom"
#         ) 
# 
# 
# 
# # combine
# costoutput_sources_1col <- costoutput_sources_n_1col + costoutput_sources_c_1col + costoutput_sources_e_1col + plot_layout(ncol = 1) +
#   plot_annotation(
#     #tag_levels = 'a', 
#     tag_levels = list(c("(a) North", "(b) Central" , "(c) East")),
#     tag_prefix = '',  
#     tag_suffix = ''
#     ) & 
#   theme(plot.tag = element_text(size = 22, face="bold"))
```



Save plot.
```{r, echo=FALSE, eval = FALSE}
# png("../figures/plot_all_outcomes_1col.png",
#     height = 30, width = 10, units = "in", res = 300)
# print(facetpanel_1col)
# dev.off()

# pdf("../figures/costs-output_sources_1col.pdf", height = 30, width = 10)
# print(costoutput_sources_1col)
# dev.off()
```



Plot again, but include output mean (without the max and min) this time.
For output, a +ve value means MIT wins.
For costs, a -ve value means MIT wins (green = MIT wins, orange = MIT loses) 
```{r}
# # First, replace output min and max with NA
# # Replace values in columns B and C with NA where A is 'output'
# sources_of_costs_2 <- sources_of_costs %>%
#   dplyr::mutate(
#     rotation_min = ifelse(source_of_cost_difference == "output", NA, rotation_min),
#     rotation_max = ifelse(source_of_cost_difference == "output", NA, rotation_max)
#   )
# 
# 
# # Create a new column to specify colours
# sources_of_costs_2 <- sources_of_costs_2 %>%
#   dplyr::mutate(color = case_when(
#     grepl("costs", source_of_cost_difference) & rotation_mean <= 0 ~ "#009E73",  # If 'costs' and y <= 0, color green
#     grepl("costs", source_of_cost_difference) & rotation_mean > 0  ~ "#CC6677",    # If 'costs' and y > 0, color red
#     grepl("output", source_of_cost_difference) & rotation_mean < 0 ~ "#CC6677",    # If 'output' and y < 0, color red
#     grepl("output", source_of_cost_difference) & rotation_mean >= 0 ~ "#009E73", # If 'output' and y >= 0, color green
#     TRUE ~ NA_character_                # Default case, if no condition is met
#   ))
# 
# 
# 
# # Create a new column for shape based on the conditions
# sources_of_costs_2 <- sources_of_costs_2 %>%
#   dplyr::mutate(pointshape = case_when(
#     source_of_cost_difference == 'output' ~ 3,  # Cross symbol
#     grepl('cost', source_of_cost_difference) ~ 16, # solid circle
#     TRUE ~ 1  # Default to a Circle symbol for other cases
#   ))
# 
# 
# 
# ggplot(sources_of_costs_2 %>% dplyr::filter(region == "north"),
#        aes(x = source_of_cost_difference, y = rotation_mean)) +
#   geom_point(aes(shape = factor(pointshape), color = color)) +  # Shape and color aesthetics
#   # geom_pointrange(aes(ymin = rotation_min, ymax = rotation_max, color = color), 
#   #                 position=position_dodge(width=0.3),
#   #                 size = 0.9) +
#   scale_shape_manual(values = c(3, 16)) +  # Cross (3) and Circle (1)
#   scale_color_identity() +  # Use the color column directly
#   facet_grid(soiltype ~ initDR, 
#              labeller = labeller(soiltype = soiltype.labs, initDR = denres.labs)#,
#              #scales="free_y"
#              ) 
# 
```
